% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assess_convergence.R
\name{assess_convergence}
\alias{assess_convergence}
\title{Assessing convergence of Metropolis-Hastings algorithm}
\usage{
assess_convergence(model_fit, items = NULL)
}
\arguments{
\item{model_fit}{A fitted model object of class \code{BayesMallows}, obtained
with \code{\link{compute_mallows}}.}

\item{items}{The items to study in the diagnostic plot for \code{rho}. A
vector of indices. If NULL, five items are selected randomly.}
}
\value{
A list containing two plots, one for alpha and one for rho.
}
\description{
Assessing convergence of Metropolis-Hastings algorithm
}
\seealso{
\code{\link{compute_mallows}}, \code{\link{plot.BayesMallows}}
}
