% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assign_cluster.R
\name{assign_cluster}
\alias{assign_cluster}
\title{Assign Assessors to Clusters}
\usage{
assign_cluster(model_fit, burnin, soft = TRUE, expand = FALSE)
}
\arguments{
\item{model_fit}{An object of type \code{BayesMallows}, returned from
\code{\link{compute_mallows}}.}

\item{burnin}{A numeric value specifying the number of iterations
to discard as burn-in. See \code{\link{assess_convergence}}.}

\item{soft}{A logical specifying whether to perform soft or
hard clustering. If \code{soft=TRUE}, all cluster probabilities
are returned, whereas if \code{soft=FALSE}, only the maximum a
posterior (MAP) cluster probability is returned, per assessor. In the
case of a tie between two or more cluster assignments, a random cluster
is taken as MAP estimate.}

\item{expand}{A logical specifying whether or not to expand the rowset
of each assessor to also include clusters for which the assessor has
0 a posterior assignment probability. Only used when \code{soft = TRUE}. Defaults
to \code{FALSE}.}
}
\value{
A dataframe.
}
\description{
Assign Assessors to Clusters
}
