% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{asymptotic_partition_function}
\alias{asymptotic_partition_function}
\title{Asymptotic Approximation of Partition Function}
\usage{
asymptotic_partition_function(alpha_grid, metric, K, n_iterations, n_items)
}
\arguments{
\item{alpha_grid}{A numeric vector of alpha values.}

\item{metric}{One of \code{"footrule"} and \code{"spearman"}.}

\item{K}{Integer.}

\item{n_iterations}{Integer specifying the number of iterations.}

\item{n_items}{Integer specifying the number of items.}
}
\value{
A vector, containing the partition function at each value of alpha.
}
\description{
Compute the asymptotic approximation of the logarithm of the partition function,
using the iteration algorithm of \insertCite{mukherjee2016;textual}{BayesMallows}.
}
\references{
\insertAllCited{}
}
\keyword{internal}
