% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BayesMallows.R
\docType{package}
\name{BayesMallows}
\alias{BayesMallows}
\alias{BayesMallows-package}
\title{BayesMallows: Bayesian Preference Learning with the Mallows Rank Model.}
\description{
The BayesMallows package provides functionality for fully
  Bayesian analysis of preference or rank data. To get started, look at
  \code{\link{compute_mallows}}.
}
