% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{beach_preferences}
\alias{beach_preferences}
\title{Pairwise preference data for beaches}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 1442 rows and 3 columns.}
\usage{
beach_preferences
}
\description{
Pairwise preference data for beaches
}
\references{
\insertRef{vitelli2018}{BayesMallows}
}
\keyword{datasets}
