% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bernoulli_data}
\alias{bernoulli_data}
\title{Simulated intransitive pairwise preferences}
\format{
An object of class \code{matrix} (inherits from \code{array}) with 120 rows and 3 columns.
}
\usage{
bernoulli_data
}
\description{
Simulated dataset based on the \link{potato_visual} data. Based on
the rankings in \link{potato_visual}, a complete set of pairwise preferences
were generated. Next, each pair of preferences was reversed with
probability 0.1, following the Bernoulli error model of
in \insertCite{crispino2019}{BayesMallows}.
}
\seealso{
Other datasets: 
\code{\link{beach_preferences}},
\code{\link{cluster_data}},
\code{\link{potato_true_ranking}},
\code{\link{potato_visual}},
\code{\link{potato_weighing}},
\code{\link{sushi_rankings}}
}
\concept{datasets}
\keyword{datasets}
