% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cluster_data}
\alias{cluster_data}
\title{Simulated clustering data}
\format{
An object of class \code{matrix} (inherits from \code{array}) with 60 rows and 5 columns.
}
\usage{
cluster_data
}
\description{
Simulated dataset of 60 complete rankings of five items, with three
different clusters.
}
\seealso{
Other datasets: 
\code{\link{beach_preferences}},
\code{\link{bernoulli_data}},
\code{\link{potato_true_ranking}},
\code{\link{potato_visual}},
\code{\link{potato_weighing}},
\code{\link{sounds}},
\code{\link{sushi_rankings}}
}
\concept{datasets}
\keyword{datasets}
