% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_consensus.R
\name{compute_consensus.consensus_SMCMallows}
\alias{compute_consensus.consensus_SMCMallows}
\title{Compute Consensus Ranking}
\usage{
\method{compute_consensus}{consensus_SMCMallows}(model_fit, type, burnin, ...)
}
\arguments{
\item{model_fit}{An object returned from \code{\link{compute_mallows}}.}

\item{type}{Character string specifying which consensus to compute. Either
\code{"CP"} or \code{"MAP"}. Defaults to \code{"CP"}.}

\item{burnin}{A numeric value specifying the number of iterations
to discard as burn-in. Defaults to \code{model_fit$burnin}, and must be
provided if \code{model_fit$burnin} does not exist. See \code{\link{assess_convergence}}.}

\item{...}{other arguments passed to methods.}
}
\description{
Compute the consensus ranking using either cumulative probability (CP) or maximum a posteriori (MAP) consensus
\insertCite{vitelli2018}{BayesMallows}. For mixture models, the
consensus is given for each mixture.
}
\seealso{
Other posteriors quantities: 
\code{\link{assign_cluster}()},
\code{\link{compute_consensus.BayesMallows}()},
\code{\link{compute_consensus}()},
\code{\link{compute_posterior_intervals.BayesMallows}()},
\code{\link{compute_posterior_intervals.SMCMallows}()},
\code{\link{compute_posterior_intervals_alpha}()},
\code{\link{compute_posterior_intervals_rho}()},
\code{\link{compute_posterior_intervals}()},
\code{\link{compute_rho_consensus}()},
\code{\link{heat_plot}()},
\code{\link{plot.BayesMallows}()},
\code{\link{plot.SMCMallows}()},
\code{\link{plot_elbow}()},
\code{\link{plot_top_k}()},
\code{\link{predict_top_k}()},
\code{\link{print.BayesMallowsMixtures}()},
\code{\link{print.BayesMallows}()}
}
\author{
Anja Stein
}
\concept{posteriors quantities}
