% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{compute_importance_sampling_estimate}
\alias{compute_importance_sampling_estimate}
\title{Mainly internal function which will be used to precompute the
partition function for footrule and Spearman distance. Can
perhaps be called withing the package if the users wishes to
use a number of items for which the partition function has not
been precomputed.}
\usage{
compute_importance_sampling_estimate(alpha_vector, n, proposal_distribution,
  target_distribution, nmc)
}
\description{
Mainly internal function which will be used to precompute the
partition function for footrule and Spearman distance. Can
perhaps be called withing the package if the users wishes to
use a number of items for which the partition function has not
been precomputed.
}
