% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_consensus.R
\name{compute_map_consensus}
\alias{compute_map_consensus}
\title{Compute MAP Consensus Ranking}
\usage{
compute_map_consensus(model_fit, burnin = model_fit$burnin)
}
\arguments{
\item{model_fit}{An object returned from \code{\link{compute_mallows}}.}

\item{burnin}{A numeric value specifying the number of iterations
to discard as burn-in. Defaults to \code{model_fit$burnin}, and must be
provided if \code{model_fit$burnin} does not exist. See \code{\link{assess_convergence}}.}
}
\description{
Compute the maximum a posterior (MAP) consensus ranking
\insertCite{vitelli2018}{BayesMallows}. For mixture models, the
consensus is given for each mixture.
}
\details{
This function is deprecated. Please use \code{\link{compute_consensus}} instead, with
argument \code{type = "MAP"}.
}
\references{
\insertAllCited{}
}
