% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_posterior.R
\name{compute_posterior}
\alias{compute_posterior}
\title{Compute the posterior distribution for the Mallows model.}
\usage{
compute_posterior(R, metric)
}
\arguments{
\item{R}{A matrix of ranks.}

\item{metric}{The metric to use.}
}
\value{
TBD
}
\description{
Compute the posterior distribution for the Mallows model.
}
\examples{
compute_posterior(potato_weighing, "footrule")
}
