% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_posterior_intervals.R
\name{compute_posterior_intervals.SMCMallows}
\alias{compute_posterior_intervals.SMCMallows}
\title{Compute posterior intervals}
\usage{
\method{compute_posterior_intervals}{SMCMallows}(
  model_fit,
  parameter = "alpha",
  level = 0.95,
  decimals = 3L,
  ...
)
}
\arguments{
\item{model_fit}{An object of class \code{SMCMallows}, returned from
\code{\link{smc_mallows_new_item_rank}} or
\code{\link{smc_mallows_new_users}}.}

\item{parameter}{Character string defining which parameter to compute
posterior intervals for. One of \code{"alpha"} or \code{"rho"}.}

\item{level}{Decimal number in \eqn{[0,1]} specifying the confidence level.
Defaults to \code{0.95}.}

\item{decimals}{Integer specifying the number of decimals to include in
posterior intervals and the mean and median. Defaults to \code{3}.}

\item{...}{Other arguments. Currently not used.}
}
\description{
This function computes posterior intervals based on the set of samples at the
last timepoint of the SMC algorithm.
}
\seealso{
Other posteriors quantities: 
\code{\link{assign_cluster}()},
\code{\link{compute_consensus.BayesMallows}()},
\code{\link{compute_consensus.consensus_SMCMallows}()},
\code{\link{compute_consensus}()},
\code{\link{compute_posterior_intervals.BayesMallows}()},
\code{\link{compute_posterior_intervals}()},
\code{\link{compute_rho_consensus}()},
\code{\link{heat_plot}()},
\code{\link{plot.BayesMallows}()},
\code{\link{plot.SMCMallows}()},
\code{\link{plot_elbow}()},
\code{\link{plot_top_k}()},
\code{\link{predict_top_k}()},
\code{\link{print.BayesMallowsMixtures}()},
\code{\link{print.BayesMallows}()}
}
\concept{posteriors quantities}
