% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smc_post_processing_functions.R
\name{compute_rho_consensus}
\alias{compute_rho_consensus}
\title{Compute rho consensus}
\usage{
compute_rho_consensus(
  output,
  nmc,
  burnin,
  C,
  type,
  colnames = NULL,
  verbose = FALSE
)
}
\arguments{
\item{output}{input}

\item{nmc}{Number of Monte Carlo samples}

\item{burnin}{A numeric value specifying the number of iterations
to discard as burn-in. Defaults to \code{model_fit$burnin}, and must be
provided if \code{model_fit$burnin} does not exist. See \code{\link{assess_convergence}}.}

\item{C}{C}

\item{type}{type}

\item{colnames}{colnames}

\item{verbose}{if \code{TRUE}, prints the final output even if the function
is assigned to an object. Defaults to \code{FALSE}.}
}
\description{
MAP AND CP consensus ranking estimates
}
\author{
Anja Stein
}
