% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{correction_kernel}
\alias{correction_kernel}
\title{Correction Kernel}
\usage{
correction_kernel(observed_ranking, current_ranking, n_items)
}
\arguments{
\item{observed_ranking}{A ranking sequence vector of the observed partial
ranking (no missing values) The original incomplete partial ranking
is in the rankings data set.}

\item{current_ranking}{A ranking sequence vector of the current augmented
ranking (no missing values)}

\item{n_items}{Integer is the number of items in a ranking}
}
\value{
List containing the proposed 'corrected' augmented ranking
that is compatible with the new observed ranking for a user
}
\description{
Function to determine if the augmented ranking is compatible
with the new observed partial ranking. If it is not, the we create a new
augmentation using the random sampling approach and calculate the
augmentation probability.
}
