% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{correction_kernel_CPP}
\alias{correction_kernel_CPP}
\title{Correction Kernel}
\usage{
correction_kernel_CPP(n_items, observed_ranking, current_ranking)
}
\arguments{
\item{n_items}{Integer is the number of items in a ranking}

\item{observed_ranking}{A ranking sequence vector of the observed partial
ranking (no missing values) The original incomplete partial ranking
is in the rankings data set.}

\item{current_ranking}{A ranking sequence vector of the current augmented
ranking (no missing values)}
}
\value{
List containing the proposed 'corrected' augmented ranking
that is compatible with the new observed ranking for a user
}
\description{
Function to determine if the augmented ranking is compatible
with the new observed partial ranking. If it is not, the we create a new
augmentation using the random sampling approachand calculate the
augmentation probability.
}
