% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_initial_ranking.R
\name{create_linear_ordering}
\alias{create_linear_ordering}
\title{Create a list of linear orderings}
\usage{
create_linear_ordering(preferences)
}
\arguments{
\item{preferences}{Dataframe with columns `assessor`, `bottom_item` and
`top_item`.}
}
\value{
A list with one element per assessor.
}
\description{
Create a list of linear orderings
}
