% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_transitive_closure.R
\name{.generate_transitive_closure}
\alias{.generate_transitive_closure}
\title{Internal function for generating transitive closure}
\usage{
.generate_transitive_closure(mat)
}
\arguments{
\item{mat}{A matrix in which column 1 is the lower ranked item and column 2 is the
upper ranked item.}
}
\description{
Internal function for generating transitive closure
}
\keyword{internal}
