% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{find_pairwise_limits}
\alias{find_pairwise_limits}
\title{Find the left and right limit for proposing augmented rank.}
\usage{
find_pairwise_limits(u, ordering, current_ranking)
}
\arguments{
\item{u}{An integer specifying the element to modify. In the modified leap-and-shift
algorithm, \eqn{u} is drawn randomly from the set \eqn{\{1, \dots, n_{items}\}}.}

\item{ordering}{A vector that contains the linear ordering of elements implied by
the preferences stated by the assessor. Note that the unconstrained elements are
not supposed to be part of this vector.}

\item{current_ranking}{A vector that contains the current complete ranking for
the assessor. This correspondings to \eqn{\tilde{R}_{j}} in the modified leap-and-shift
algorithm.}
}
\value{
A vector of size 2. The first element is the left limit (\eqn{l_{j}}) and
the second element is the right limit (\eqn{r_{j}}).
}
\description{
This function implements a part of the modified leap-and-shift
algorithm for proposing augmented ranks that agree with the
transitive closure of pairwise preferences stated by the assessor. The
algorithm is described on pp. 21-22 of \insertCite{vitelli2018;textual}{BayesMallows}.
\code{find_pairwise_limits} returns the left and right limits, \eqn{l_{j}} and
\eqn{r_{j}}. The proposed new ranking of the given element is then
sampled uniformly from the set \eqn{\{l_{j} + 1, \dots, r_{j} - 1\}}, whereupon
a shift step is applied. This function is separated out mainly for testing
purposes.
}
\references{
\insertAllCited{}
}
\keyword{internal}
