% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_cardinalities.R
\docType{data}
\name{footrule_cardinalities}
\alias{footrule_cardinalities}
\alias{spearman_cardinalities}
\alias{ulam_cardinalities}
\title{Distances and cardinalities}
\format{
An object of class \code{list} of length 50.

An object of class \code{list} of length 20.

An object of class \code{list} of length 60.
}
\usage{
footrule_cardinalities

spearman_cardinalities

ulam_cardinalities
}
\description{
Lists containing distances and cardinalities for footrule,
Spearman, and Ulam distances. See \code{\link[=get_cardinalities]{get_cardinalities()}} for details.
}
\seealso{
Other partition function: 
\code{\link{estimate_partition_function}()},
\code{\link{get_cardinalities}()}
}
\concept{partition function}
\keyword{datasets}
