% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_initial_ranking.R
\name{generate_initial_ranking.BayesMallowsIntransitive}
\alias{generate_initial_ranking.BayesMallowsIntransitive}
\title{Generate initial ranking for non-transitive preferences}
\usage{
\method{generate_initial_ranking}{BayesMallowsIntransitive}(preferences, n_items, cl = NULL, ...)
}
\arguments{
\item{preferences}{A data frame with one row per pairwise comparison, and
  columns \code{assessor}, \code{top_item}, and \code{bottom_item}. Each
  column contains the following:
\itemize{
\item \code{assessor} is a numeric vector containing the assessor index, or a character
      vector containing the (unique) name of the assessor.

\item \code{bottom_item} is a numeric vector containing the index of the item that
      was disfavored in each pairwise comparison.

\item \code{top_item} is a numeric vector containing the index of the item that was
      preferred in each pairwise comparison.
}
  So if we have two assessors and five items, and assessor 1 prefers item 1
  to item 2 and item 1 to item 5, while assessor 2 prefers item 3 to item 5,
  we have the following \code{df}:
\tabular{rrr}{
\strong{assessor} \tab \strong{bottom_item} \tab \strong{top_item}\cr
1 \tab 2 \tab 1\cr
1 \tab 5 \tab 1\cr
2 \tab 5 \tab 3\cr
}}

\item{n_items}{Number of items.}

\item{cl}{Cluster. Not used.}
}
\value{
A matrix of rankings which can be given in the \code{rankings}
  argument to \code{\link{compute_mallows}}.
}
\description{
Generate initial ranking for non-transitive preferences
}
\seealso{
Other preprocessing: 
\code{\link{estimate_partition_function}()},
\code{\link{generate_constraints}()},
\code{\link{generate_initial_ranking.BayesMallowsTC}()},
\code{\link{generate_transitive_closure}()},
\code{\link{obs_freq}},
\code{\link{prepare_partition_function}()}
}
\concept{preprocessing}
