% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_transitive_closure.R
\name{generate_transitive_closure}
\alias{generate_transitive_closure}
\title{Title}
\usage{
generate_transitive_closure(df, num_items)
}
\arguments{
\item{df}{Dataframe with columns \code{Assessor}, \code{BottomItem}, and
\code{TopItem}.}

\item{num_items}{Number of items in total.}
}
\description{
Title
}
