% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{get_partition_function}
\alias{get_partition_function}
\title{Compute the logarithm of the partition function for a Mallows rank model.}
\usage{
get_partition_function(alpha, summation_sequences, metric = "footrule")
}
\arguments{
\item{alpha}{The value of the alpha parameter.}

\item{summation_sequences}{List with elements \code{distances} and
\code{cardinalities}, both of type arma::vec.}

\item{metric}{A string. Avaiable options are \code{"footrule"},
\code{"kendall"}, and \code{"spearman"}. Defaults to \code{"footrule"}.}
}
\value{
A scalar, the logarithm of the partition function.
}
\description{
Compute the logarithm of the partition function for a Mallows rank model.
}
