% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{get_partition_functions}
\alias{get_partition_functions}
\title{Compute the logarithm of the partition function for a Mallows rank model.}
\usage{
get_partition_functions(n, alpha, metric = "footrule")
}
\arguments{
\item{n}{The number of items, a positive integer.}

\item{alpha}{The value of the alpha parameter.}

\item{metric}{A string. Avaiable options are \code{"footrule"},
\code{"kendall"}, and \code{"spearman"}. Defaults to \code{"footrule"}.}
}
\value{
A scalar, the logarithm of the partition function.
}
\description{
Compute the logarithm of the partition function for a Mallows rank model.
}
