% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{get_rank_distance}
\alias{get_rank_distance}
\title{Compute the distance between two rank vectors.}
\usage{
get_rank_distance(r1, r2, metric = "footrule")
}
\arguments{
\item{r1}{A vector of ranks.}

\item{r2}{A vector of ranks.}

\item{metric}{A string. Avaiable options are \code{"footrule"},
\code{"kendall"}, and \code{"spearman"}. Defaults to \code{"footrule"}.}
}
\value{
A scalar.
}
\description{
Compute the distance between two rank vectors.
}
