% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{get_summation_distances}
\alias{get_summation_distances}
\title{Get the distances for computing the partition function given
the cardinalities.}
\usage{
get_summation_distances(n, cardinalities, metric = "footrule")
}
\arguments{
\item{metric}{A string. Avaiable options are \code{"footrule"},
\code{"kendall"}, and \code{"spearman"}. Defaults to \code{"footrule"}.}

\item{alpha}{The value of the alpha parameter.}

\item{summation_sequences}{List with elements \code{distances} and
\code{cardinalities}, both of type arma::vec.}
}
\value{
A scalar, the logarithm of the partition function.
}
\description{
Get the distances for computing the partition function given
the cardinalities.
}
