% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_transitive_closure.R
\name{get_transitive_closure.BayesMallowsData}
\alias{get_transitive_closure.BayesMallowsData}
\title{Get transitive closure}
\usage{
\method{get_transitive_closure}{BayesMallowsData}(rank_data)
}
\arguments{
\item{rank_data}{An object of class \code{"BayesMallowsData"} returned from
\code{\link{setup_rank_data}}.}
}
\value{
A dataframe with transitive closure, if these is any.
}
\description{
Get transitive closure
}
