% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_transitive_closure.R
\name{get_transitive_closure}
\alias{get_transitive_closure}
\title{Generic function for obtaining transitive closure}
\usage{
get_transitive_closure(rank_data)
}
\description{
Generic function for obtaining transitive closure
}
