% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{importance_sampling_fit}
\alias{importance_sampling_fit}
\title{Importance sampling fits in the form of regression coefficients based on
\code{log(Z) = b0 + b1 alpha + ... + b10 alpha^10}. These are to be supplied
to \code{\link{run_mcmc}} internally, when the importance sampling estimates are to
be used.}
\format{An object of class \code{list} of length 6.}
\usage{
importance_sampling_fit
}
\description{
Importance sampling fits in the form of regression coefficients based on
\code{log(Z) = b0 + b1 alpha + ... + b10 alpha^10}. These are to be supplied
to \code{\link{run_mcmc}} internally, when the importance sampling estimates are to
be used.
}
\keyword{datasets}
