% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc_likelihood.R
\name{log_lik_db}
\alias{log_lik_db}
\title{Log-likelihood evaluation for a Mallows model}
\usage{
log_lik_db(rho, alpha, metric, rankings, obs_freq)
}
\arguments{
\item{rho}{Numeric vector of length n_items. \code{rho} must be a permutation
of the first n_items integers and corresponds to the location of the
Mallows distribution. The location coincides with the mode (most probable
permutation).}

\item{alpha}{Non-negative scalar specifying the scale (precision) parameter
in the Mallows rank model with the Kendall distance.}

\item{metric}{Character string specifying the distance measure to use.
Available options are \code{"kendall"}, \code{"cayley"}, \code{"hamming"},
\code{"ulam"} for \code{n_items<=95}, \code{"footrule"} for
\code{n_items<=50} and \code{"spearman"} for \code{n_items<=14}.}

\item{rankings}{A matrix with observed rankings in each row.}

\item{obs_freq}{A vector of observation frequencies (weights) to apply to
each row in \code{rankings}. This can speed up computation if a large
number of assessors share the same rank pattern.}
}
\value{
The log-likelihood value corresponding to one or more observed
  rankings under the Mallows rank model with distance specified by the
  \code{metric} argument.
}
\description{
Compute the log-likelihood value of the Mallows model parameters
  for a ranking dataset.
}
\keyword{internal}
