% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc_likelihood.R
\name{log_lik_db_mix}
\alias{log_lik_db_mix}
\title{Log-likelihood evaluation for a Mallows mixture model}
\usage{
log_lik_db_mix(rho, alpha, weights, metric, rankings, obs_freq)
}
\arguments{
\item{rho}{A matrix of size \code{n_clusters x n_items} whose rows are
permutations of the first n_items integers corresponding to the modal
rankings of the Mallows mixture components.}

\item{alpha}{A vector of \code{n_clusters} non-negative scalars specifying
the scale (precision) parameters of the Mallows mixture components.}

\item{weights}{A vector of \code{n_clusters} non-negative scalars specifying
the mixture weights.}

\item{metric}{Character string specifying the distance measure to use.}

\item{rankings}{A matrix with observed rankings in each row.}

\item{obs_freq}{A vector of observation frequencies (weights) to apply to
each row in \code{rankings}. This can speed up computation if a large
number of assessors share the same rank pattern. If \code{NULL}, it means
that each row of \code{rankings} is multiplied by 1. If provided,
\code{obs_freq} must have the same number of elements as there are rows in
\code{rankings}, and \code{rankings} cannot be \code{NULL}.}
}
\value{
The log-likelihood value corresponding to one or more observed
  rankings under the Mallows mixture model with distance specified by the
  \code{metric} argument.
}
\description{
Compute the log-likelihood value of the Mallows mixture model
  parameters for a ranking dataset.
}
\keyword{internal}
