% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smc_uniform_functions.R
\name{metropolis_hastings_aug_ranking}
\alias{metropolis_hastings_aug_ranking}
\title{Metropolis-Hastings Augmented Ranking}
\usage{
metropolis_hastings_aug_ranking(R_curr, R_obs, alpha, rho, n_items, metric)
}
\arguments{
\item{R_curr}{A ranking sequence vector of the current augmented ranking (no missing values)}

\item{R_obs}{A ranking sequence vector of the observed partial ranking (no missing values) The original incomplete partial ranking is in the rankings data set.}

\item{alpha}{Numeric value og the scale parameter}

\item{rho}{A ranking sequence vector}

\item{n_items}{Integer is the number of items in a ranking}

\item{metric}{A character string specifying the distance metric to use in the
Bayesian Mallows Model. Available options are \code{"footrule"},
\code{"spearman"}, \code{"cayley"}, \code{"hamming"}, \code{"kendall"}, and
\code{"ulam"}.}
}
\value{
R_curr or R_obs A ranking sequence vector representing proposed augmented ranking for next
        iteration of MCMC chain
}
\description{
Function to perform Metropolis-Hastings for new augmented ranking
}
