% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test.R
\name{metropolis_hastings_aug_ranking_wrapper}
\alias{metropolis_hastings_aug_ranking_wrapper}
\title{Title}
\usage{
metropolis_hastings_aug_ranking_wrapper(
  alpha,
  rho,
  n_items,
  partial_ranking,
  current_ranking,
  metric,
  seed = 123L
)
}
\arguments{
\item{alpha}{aa}

\item{rho}{aa}

\item{n_items}{aa}

\item{partial_ranking}{aa}

\item{current_ranking}{aa}

\item{metric}{aa}

\item{seed}{aaa}
}
\value{

}
\description{
Title
}
