% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_processing_functions.R
\name{plot_alpha_posterior}
\alias{plot_alpha_posterior}
\title{Plot Alpha Posterior}
\usage{
plot_alpha_posterior(output, nmc, burnin)
}
\arguments{
\item{output}{output}

\item{nmc}{nmc}

\item{burnin}{burn-in}
}
\description{
posterior for alpha
}
