% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_elbow.R
\name{plot_elbow}
\alias{plot_elbow}
\title{Plot Within-Cluster Sum of Distances}
\usage{
plot_elbow(..., burnin = NULL)
}
\arguments{
\item{...}{One or more objects returned from \code{\link{compute_mallows}},
separated by comma, or a list of such objects. Typically, each object
has been run with a different number of mixtures, as specified in the
\code{n_clusters} argument to \code{\link{compute_mallows}}.}

\item{burnin}{The number of iterations to discard as burnin. The following
options are possible:
\itemize{
\item A numeric vector with one element for each model object provided in
\code{...}.

\item A single number. In this case, the burnin for each model provided in
\code{...} is taken to be equal to this single number.

\item If the argument to \code{...} is a list of models, and this list
has an element named \code{burnin}, then this is the default value of \code{burnin}.
The element can either be a vector or a single number, as described in the points
above.

\item If the arguments to \code{...} are a set of models, and each
models has a \code{burnin} element set, then this is the default value of
\code{burnin}.
}}
}
\value{
A boxplot with the number of clusters on the horizontal axis and the
with-cluster sum of distances on the vertical axis.
}
\description{
Plot the within-cluster sum of distances from the corresponding cluster
consensus for different number of clusters. This function is useful for
selecting the number of mixture.
}
\seealso{
\code{\link{compute_mallows}}
}
