% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smc_post_processing_functions.R
\name{plot_rho_heatplot}
\alias{plot_rho_heatplot}
\title{Plot rho heatplot}
\usage{
plot_rho_heatplot(output, nmc, burnin, n_items, rho)
}
\arguments{
\item{output}{input}

\item{nmc}{Number of Monte Carlo samples}

\item{burnin}{A numeric value specifying the number of iterations
to discard as burn-in. Defaults to \code{model_fit$burnin}, and must be
provided if \code{model_fit$burnin} does not exist. See \code{\link{assess_convergence}}}

\item{n_items}{Integer is the number of items in the consensus ranking}

\item{rho}{Numeric vector specifying the current consensus ranking}
}
\value{
A heatplot
}
\description{
Plot rho heatplot
}
\author{
Anja Stein
}
