% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.SMCMallows.R
\name{plot_rho_posterior}
\alias{plot_rho_posterior}
\title{Plot the posterior for rho for each item}
\usage{
plot_rho_posterior(output, nmc, burnin, C, colnames = NULL, items = NULL)
}
\arguments{
\item{output}{input}

\item{nmc}{Number of Monte Carlo samples}

\item{burnin}{A numeric value specifying the number of iterations
to discard as burn-in. Defaults to \code{model_fit$burnin}, and must be
provided if \code{model_fit$burnin} does not exist. See \code{\link{assess_convergence}}}

\item{C}{Number of cluster}

\item{colnames}{A vector of item names. If NULL, we generate generic names for the items in the ranking.}

\item{items}{Either a vector of item names, or a
vector of indices. If NULL, five items are selected randomly.}
}
\description{
Plot the posterior for rho for each item
}
