% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smc_post_processing_functions.R
\name{plot_rho_trace}
\alias{plot_rho_trace}
\title{Plot rho trace}
\usage{
plot_rho_trace(output, nmc)
}
\arguments{
\item{output}{output}

\item{nmc}{nmc}
}
\description{
Plot rho trace
}
\author{
Anja Stein
}
