% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_top_k.R
\name{plot_top_k}
\alias{plot_top_k}
\title{Find the top-k probabilities}
\usage{
plot_top_k(model_fit, burnin, k = 3, rel_widths = c(rep(1,
  model_fit$n_clusters), 10))
}
\arguments{
\item{model_fit}{An object of type \code{BayesMallows}, returned from
\code{\link{compute_mallows}}.}

\item{burnin}{Number of iterations to discard as burn-in.}

\item{k}{The k in top-k}

\item{rel_widths}{The relative widths of the plots of \code{rho} per cluster
and the plot of assessors, respectively. This argument is passed on to
\code{cowplot::plot_grid}.}
}
\description{
Find the top-k probabilities
}
