% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{potato_partial}
\alias{potato_partial}
\title{Updated partial rankings of potatoes}
\format{
An object of class \code{array} of dimension 12 x 20 x 11.
}
\usage{
potato_partial
}
\description{
Simulated dataset based on \code{\link{potato_visual}}, in which 50 \% of the
potatoes are initially unranked. Through ten subsequent timesteps, more
ranks are added each time, given complete rankings at time 11.
}
\references{
\insertRef{liu2019}{BayesMallows}
}
\seealso{
Other datasets: 
\code{\link{beach_preferences}},
\code{\link{potato_true_ranking}},
\code{\link{potato_visual}},
\code{\link{potato_weighing}},
\code{\link{sample_dataset}},
\code{\link{sushi_rankings}}
}
\concept{datasets}
\keyword{datasets}
