% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{potato_true_ranking}
\alias{potato_true_ranking}
\title{True ranking of the weights of 20 potatoes.}
\format{An integer vector with 20 elements}
\usage{
potato_true_ranking
}
\description{
True ranking of the weights of 20 potatoes.
}
\references{
Insert reference to review paper where the potato data is presented.
}
\keyword{datasets}
