% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{potato_weighing}
\alias{potato_weighing}
\title{Result of ranking potatoes by weight, where the assessors were allowed to lift the potatoes. 12 assessors ranked 20 potatoes.}
\format{An integer matrix with 12 rows and 20 columns.}
\usage{
potato_weighing
}
\description{
Result of ranking potatoes by weight, where the assessors were allowed to lift the potatoes. 12 assessors ranked 20 potatoes.
}
\references{
Insert reference to review paper where the potato data is presented.
}
\keyword{datasets}
