% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{prepare_partition_function}
\alias{prepare_partition_function}
\title{Prepare partition functions}
\usage{
prepare_partition_function(logz_estimate = NULL, metric, n_items)
}
\arguments{
\item{logz_estimate}{Optional argument containing the result of calling
\code{\link{estimate_partition_function}}.}

\item{metric}{Metric to be used.}

\item{n_items}{Number of items.}
}
\value{
List with two elements, \code{cardinalities} and \code{logz_estimate},
one of which is \code{NULL} and the other of which contains partition function
estimates.
}
\description{
Utility function for estimating partition function of the Mallows model.
}
\seealso{
Other preprocessing: 
\code{\link{estimate_partition_function}()},
\code{\link{generate_constraints}()},
\code{\link{generate_initial_ranking.BayesMallowsIntransitive}()},
\code{\link{generate_initial_ranking.BayesMallowsTC}()},
\code{\link{generate_transitive_closure}()},
\code{\link{obs_freq}}
}
\concept{preprocessing}
