% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.BayesMallows}
\alias{print.BayesMallows}
\alias{print.BayesMallowsMixtures}
\alias{print.SMCMallows}
\title{Print Method for BayesMallows Objects}
\usage{
\method{print}{BayesMallows}(x, ...)

\method{print}{BayesMallowsMixtures}(x, ...)

\method{print}{SMCMallows}(x, ...)
}
\arguments{
\item{x}{An object of type \code{BayesMallows}, returned from
\code{\link[=compute_mallows]{compute_mallows()}}.}

\item{...}{Other arguments passed to \code{print} (not used).}
}
\description{
The default print method for a \code{BayesMallows} object.
}
\seealso{
Other posterior quantities: 
\code{\link{assign_cluster}()},
\code{\link{compute_consensus}()},
\code{\link{compute_posterior_intervals}()},
\code{\link{heat_plot}()},
\code{\link{plot.BayesMallows}()},
\code{\link{plot.SMCMallows}()},
\code{\link{plot_elbow}()},
\code{\link{plot_top_k}()},
\code{\link{predict_top_k}()}
}
\concept{posterior quantities}
