% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rank_distance.R
\name{rank_distance}
\alias{rank_distance}
\title{Distance between two rankings}
\usage{
rank_distance(r1, r2, metric)
}
\arguments{
\item{r1}{A vector of ranks.}

\item{r2}{A vector of ranks.}

\item{metric}{Character string specifying the distance measure to use. Available options are \code{"kendall"}, \code{"cayley"}, \code{"hamming"}, \code{"ulam"}, \code{"footrule"} and \code{"spearman"}.}
}
\value{
A scalar equal to the distance between the two ranking sequences according to the given \code{metric}.
}
\description{
Compute the distance between two rankings according to several metrics.
}
\details{
Note that the Spearman distance is the squared L2 norm, whereas
the footrule distance is the L1 norm.

The Ulam distance uses the SUBSET library developed by John Burkardt, available at http://people.sc.fsu.edu/~jburkardt/cpp_src/subset/subset.html.

The implementation of Cayley distance is based on a \code{C++} translation of \code{Rankcluster::distCayley} \insertCite{Grimonprez2016}{BayesMallows}.
}
\examples{
rank_distance(1:5,5:1,metric="kendall")
rank_distance(c(2,4,3,6,1,7,5),c(3,5,4,7,6,2,1),metric="cayley")
rank_distance(c(4,2,3,1),c(3,4,1,2),metric="hamming")
rank_distance(c(1,3,5,7,9,8,6,4,2),c(1,2,3,4,9,8,7,6,5),"ulam")
rank_distance(c(8,7,1,2,6,5,3,4),c(1,2,8,7,3,4,6,5),"footrule")
rank_distance(c(1,6,2,5,3,4),c(4,3,5,2,6,1),"spearman")
}
\references{
\insertAllCited{}
}
