% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rank_distance_vec.R
\name{rank_distance_vec}
\alias{rank_distance_vec}
\title{Distance between a set of rankings and a given rank sequence}
\usage{
rank_distance_vec(rankings, rho, metric)
}
\arguments{
\item{rankings}{A matrix of size \eqn{N }\eqn{\times}{x}\eqn{ n_items} of rankings in each row.}

\item{rho}{A ranking sequence.}

\item{metric}{Character string specifying the distance measure to use. Available options are \code{"kendall"}, \code{"cayley"}, \code{"hamming"}, \code{"ulam"}, \code{"footrule"} and \code{"spearman"}.}
}
\value{
A vector of distances according to the given \code{metric}.
}
\description{
Compute the distance between a matrix of rankings and a rank sequence.
}
\examples{
data0=sample_mallows(rho0=1:10,alpha=20,n_samples=1000)

rank_distance_vec(rankings=data0,rho=1:10,metric="kendall")
}
\references{
\insertAllCited{}
}
