% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rtruncbeta}
\alias{rtruncbeta}
\title{Random numbers from the truncated beta distribution.}
\usage{
rtruncbeta(shape1, shape2, trunc = 1)
}
\arguments{
\item{shape1}{First shape parameter.}

\item{shape2}{Second shape parameter.}

\item{trunc}{Where to truncate.}
}
\description{
Random numbers from the truncated beta distribution.
}
