% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{run_mcmc}
\alias{run_mcmc}
\title{Worker function for computing the posterior distribtuion.}
\usage{
run_mcmc(R, cardinalities, metric = "footrule", nmc = 100L, L = 1L,
  sd_alpha = 0.5, lambda = 0.1, alpha_init = 5)
}
\arguments{
\item{R}{A set of complete rankings, with one sample per column.
With N samples and n items, R is n x N.}

\item{cardinalities}{Used when metric equals \code{"footrule"} or
\code{"spearman"} for computing the partition function.
TODO: Make this argument optional.}

\item{metric}{The distance metric to use.}

\item{nmc}{Number of Monte Carlo samples.}

\item{L}{Leap-and-shift step size.}
}
\description{
Worker function for computing the posterior distribtuion.
}
