% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_mallows.R
\name{sample_mallows}
\alias{sample_mallows}
\title{Random Samples from the Mallows Rank Model}
\usage{
sample_mallows(rho0, alpha0, n_samples, burnin, thinning, leap_size = 1,
  metric = "footrule", diagnostic = FALSE, items_to_plot = NULL)
}
\arguments{
\item{rho0}{Vector specifying the latent consensus ranking.}

\item{alpha0}{Scalar specifying the scale parameter.}

\item{n_samples}{Integer specyfing the number of random samples to generate.}

\item{burnin}{Integer specifying the number of iterations to discard as burn-in.}

\item{thinning}{Integer specifying the number of MCMC iterations to perform
between each time a random rank vector is sampled.}

\item{leap_size}{Integer specifying the step size of the leap-and-shift proposal distribution.}

\item{metric}{Character string specifying the distance measure to use. Available
options are \code{"footrule"} (default), \code{"spearman"}, \code{"cayley"}, and
\code{"kendall"}. For sampling from the Mallows model with Cayley and Kendall distances
the \code{PerMallows} package \insertCite{irurozki2016}{BayesMallows} can also be used.}

\item{diagnostic}{Logical specifying whether to output convergence diagnostics. If \code{TRUE},
a diagnostic plot is returned together with the samples.}

\item{items_to_plot}{Integer vector used if \code{diagnostic = TRUE}, in order to
specify the items to plot in the diagnostic output. If not provided, 5 items are picked
at random.}
}
\description{
Generate random samples from the Mallows
Rank Model \insertCite{mallows1957}{BayesMallows} with consensus ranking \eqn{\rho}
and scale parameter \eqn{\alpha}. The samples are obtained by running the
Metropolis-Hastings algorithm described in Appendix C of
\insertCite{vitelli2018;textual}{BayesMallows}.
}
\examples{
# Example coming.
# Number of items
n_items <- 15
# Set the consensus ranking
rho0 <- seq(from = 1, to = n_items, by = 1)
# Set the scale
alpha0 <- 5
# Number of samples we want
n_samples <- 2
# Burnin
burnin <- 1000
# Thinning
thinning <- 100

sample_mallows(rho0, alpha0, n_samples, burnin, thinning,
               diagnostic = TRUE)
}
\references{
\insertAllCited{}
}
