% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_prior.R
\name{sample_prior}
\alias{sample_prior}
\title{Sample from prior distribution}
\usage{
sample_prior(n, n_items, priors = set_priors())
}
\arguments{
\item{n}{An integer specifying the number of samples to take.}

\item{n_items}{An integer specifying the number of items to be ranked.}

\item{priors}{An object of class "BayesMallowsPriors" returned from
\code{\link[=set_priors]{set_priors()}}.}
}
\value{
An object of class "BayesMallowsPriorSample", containing \code{n}
independent samples of \eqn{\alpha} and \eqn{\rho}.
}
\description{
Function to obtain samples from the prior distributions of the Bayesian
Mallows model. Intended to be given to \code{\link[=update_mallows]{update_mallows()}}.
}
\examples{
prior_samples <- sample_prior(10000, ncol(sushi_rankings))
model1 <- update_mallows(
  model = prior_samples,
  new_data = setup_rank_data(sushi_rankings[1:500, ]),
  smc_options = set_smc_options(mcmc_steps = 20))
plot(model1)

model2 <- update_mallows(
  model = model1,
  new_data = setup_rank_data(sushi_rankings[501:1000, ]))
plot(model2)

model3 <- update_mallows(
  model = model2,
  new_data = setup_rank_data(sushi_rankings[1001:1500, ]))
plot(model3)
}
\seealso{
Other modeling: 
\code{\link{compute_mallows_mixtures}()},
\code{\link{compute_mallows}()},
\code{\link{update_mallows}()}
}
\concept{modeling}
