% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_smc_options.R
\name{set_smc_options}
\alias{set_smc_options}
\title{Set SMC compute options}
\usage{
set_smc_options(n_particles = 1000, mcmc_steps = 5)
}
\arguments{
\item{n_particles}{Integer specifying the number of particles.}

\item{mcmc_steps}{Number of MCMC steps to be applied in the resample-move
step.}
}
\value{
An object of class "SMCOptions".
}
\description{
Sets the SMC compute options to be used in \code{\link[=update_mallows.BayesMallows]{update_mallows.BayesMallows()}}.
}
\seealso{
Other preprocessing: 
\code{\link{get_transitive_closure}()},
\code{\link{set_compute_options}()},
\code{\link{set_initial_values}()},
\code{\link{set_model_options}()},
\code{\link{set_priors}()},
\code{\link{setup_rank_data}()}
}
\concept{preprocessing}
