% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smc_mallows_update.R
\name{smc_mallows_update}
\alias{smc_mallows_update}
\title{Generic function for updating SMC Mallows models}
\usage{
smc_mallows_update(model, rankings, ...)
}
\arguments{
\item{model}{A model object.}

\item{rankings}{A matrix with rankings.}

\item{...}{Other optional arguments.}
}
\value{
An updated model.
}
\description{
Generic function for updating SMC Mallows models
}
