% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smc_mallows_update.R
\name{smc_mallows_update.SMCMallowsNewUsers}
\alias{smc_mallows_update.SMCMallowsNewUsers}
\title{Update an SMC Mallows model with new users}
\usage{
\method{smc_mallows_update}{SMCMallowsNewUsers}(
  model,
  rankings,
  timesteps = 1,
  num_new_obs = nrow(rankings),
  n_particles = model$n_particles,
  verbose = model$verbose,
  ...
)
}
\arguments{
\item{model}{Object of class \code{c("SMCMallowsNewUsers","SMCMallows")}
returned from \code{\link{smc_mallows_new_users}}.}

\item{rankings}{Matrix containing the new set of observed rankings of size
n_assessors by n_items.}

\item{timesteps}{Integer specifying the number of timesteps. Defaults to 1.}

\item{num_new_obs}{Integer specifying the number of new observations.
Defaults to \code{nrow(rankings)}.}

\item{n_particles}{Integer specifying the number of particles. Defaults to
the number of particles used when computing \code{model}.}

\item{verbose}{Logical specifying whether to print out the progress of the
SMC-Mallows algorithm. Defaults to \code{FALSE}.}

\item{...}{Optional additional arguments. Currently not used.}
}
\value{
An updated model, of class
  \code{c("SMCMallowsNewUsers","SMCMallows")}.
}
\description{
Update an SMC Mallows model with new users
}
\examples{
# Assume our first dataset contains 100 preferences, observed in 10 batches
data_batch1 <- sushi_rankings[1:100, ]

# Calculating rho and alpha samples
mod1 <- smc_mallows_new_users(
  rankings = data_batch1,
  n_particles = 100,
  timesteps = 10,
  mcmc_kernel_app = 10,
  num_new_obs = 10
)

# We can plot the posterior of alpha
plot(mod1)

# Next, assume a new batch arrives, with 100 new preferences
data_batch2 <- sushi_rankings[101:200, ]

# We update the model in a single timestep. By default, the settings from the
# call to smc_mallows_new_users() is retained.
mod2 <- smc_mallows_update(
  model = mod1,
  rankings = data_batch2
)

plot(mod2)
}
\seealso{
Other modeling: 
\code{\link{compute_mallows_mixtures}()},
\code{\link{compute_mallows}()},
\code{\link{smc_mallows_new_item_rank}()},
\code{\link{smc_mallows_new_users}()}
}
\concept{modeling}
