% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smc_post_processing_functions.R
\name{smc_processing}
\alias{smc_processing}
\title{SMC Processing}
\usage{
smc_processing(output, colnames = NULL)
}
\arguments{
\item{output}{input}

\item{colnames}{colnames}
}
\description{
SMC Processing
}
\author{
Anja Stein
}
