% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sounds}
\alias{sounds}
\title{Sounds data}
\format{
An object of class \code{data.frame} with 1380 rows and 3 columns.
}
\usage{
sounds
}
\description{
Data from an experiment in which 46 individuals compared 12 different sounds
\insertCite{barrett2018}{BayesMallows}. Each assessor was asked multiple
times to compare a pair of two sounds, indicating which of the sounds sounded
the most like it was human generated. The pairwise preference for each
assessor are in general non-transitive. These data inspired the Mallows model
for non-transitive pairwise preferences developed by
\insertCite{crispino2019}{BayesMallows}.
}
\references{
\insertAllCited{}
}
\seealso{
Other datasets: 
\code{\link{beach_preferences}},
\code{\link{bernoulli_data}},
\code{\link{cluster_data}},
\code{\link{potato_true_ranking}},
\code{\link{potato_visual}},
\code{\link{potato_weighing}},
\code{\link{sushi_rankings}}
}
\concept{datasets}
\keyword{datasets}
