% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_cardinalities.R
\docType{data}
\name{spearman_cardinalities}
\alias{spearman_cardinalities}
\title{Distances and cardinalities}
\format{
An object of class \code{list} of length 20.
}
\usage{
spearman_cardinalities
}
\description{
List containing distances and cardinalities for Spearman
distance. See \code{\link[=get_cardinalities]{get_cardinalities()}} for details.
}
\seealso{
Other partition function: 
\code{\link{estimate_partition_function}()},
\code{\link{footrule_cardinalities}},
\code{\link{get_cardinalities}()},
\code{\link{ulam_cardinalities}}
}
\concept{partition function}
\keyword{datasets}
