% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sushi_rankings}
\alias{sushi_rankings}
\title{Sushi preferences}
\format{An object of class \code{matrix} with 5000 rows and 10 columns.}
\usage{
sushi_rankings
}
\description{
Sushi preferences
}
\references{
\insertRef{vitelli2018}{BayesMallows}
}
\keyword{datasets}
