% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_mcmc.R
\name{tidy_mcmc}
\alias{tidy_mcmc}
\title{Internal Function for Tidying MCMC Output}
\usage{
tidy_mcmc(fit, tidy_rho = TRUE, tidy_alpha = TRUE,
  tidy_cluster_assignment = TRUE, tidy_cluster_probabilities = TRUE,
  tidy_wcd = TRUE, tidy_augmented_data = TRUE,
  tidy_augmentation_acceptance = TRUE)
}
\arguments{
\item{fit}{A fitted object, returned from \code{BayesMallows:::run_mcmc}.}

\item{tidy_rho}{Logical specifying whether or not to tidy the output for \code{rho}.
Defaults to \code{TRUE}.}

\item{tidy_alpha}{Logical specifying whether or not to tidy the output for \code{alpha}.
Defaults to \code{TRUE}.}

\item{tidy_cluster_assignment}{Logical specifying whether or not to tidy the output for
cluster assignments. Defaults to \code{TRUE}.}

\item{tidy_cluster_probabilities}{Logical specifying whether or not to tidy the output for
cluster probabilities. Defaults to \code{TRUE}.}

\item{tidy_wcd}{Logical specifying whether or not to tidy the output for
within-cluster distances. Defaults to \code{TRUE}.}

\item{tidy_augmented_data}{Logical specifying whether or not to tidy the output for
augmented data. Defaults to \code{TRUE}.}

\item{tidy_augmentation_acceptance}{Logical specifying whether or not to tidy the output for
augmentation acceptance. Defaults to \code{TRUE}.}
}
\description{
Internal Function for Tidying MCMC Output
}
\examples{
# Run compute_mallows with the option skip_postprocessing=TRUE.
n_clusters <- 1:10
library(purrr)
models <- n_clusters \%>\%
  map(~ compute_mallows(potato_visual, n_clusters = .x,
                        include_wcd = TRUE, skip_postprocessing = TRUE))

# We are now not able to compute an elbow plot
\dontrun{
  plot_elbow(models, burnin = 500)
}

# We need to tidy the parts that are used by plot_elbow.
tidy_models <- models \%>\%
  map(~ tidy_mcmc(.x,
                  tidy_rho = FALSE,
                  tidy_alpha = FALSE,
                  tidy_cluster_assignment = FALSE,
                  tidy_cluster_probabilities = FALSE,
                  tidy_wcd = TRUE,
                  tidy_augmented_data = TRUE,
                  tidy_augmentation_acceptance = TRUE)
      )
plot_elbow(tidy_models, burnin = 500)
}
\keyword{internal}
