% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_cardinalities.R
\docType{data}
\name{ulam_cardinalities}
\alias{ulam_cardinalities}
\title{Distances and cardinalities}
\format{
An object of class \code{list} of length 60.
}
\usage{
ulam_cardinalities
}
\description{
List containing distances and cardinalities for Ulam
distance. See \code{\link[=get_cardinalities]{get_cardinalities()}} for details.
}
\seealso{
Other partition function: 
\code{\link{estimate_partition_function}()},
\code{\link{footrule_cardinalities}},
\code{\link{get_cardinalities}()},
\code{\link{spearman_cardinalities}}
}
\concept{partition function}
\keyword{datasets}
