% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_mallows.R
\name{update_mallows}
\alias{update_mallows}
\title{Generic function for updating Mallows models using SMC}
\usage{
update_mallows(model, new_rankings, ...)
}
\arguments{
\item{model}{A model object.}

\item{new_rankings}{A matrix with rankings.}

\item{...}{Other optional arguments.}
}
\value{
An updated model.
}
\description{
Generic function for updating Mallows models using SMC
}
