% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{validate_permutation}
\alias{validate_permutation}
\title{Check if a vector is a permutation}
\usage{
validate_permutation(vec)
}
\arguments{
\item{vec}{a vector}
}
\value{
TRUE if vec is a permutation
}
\description{
Check if a vector is a permutation
}
\keyword{internal}
