% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{validate_rank_ordering}
\alias{validate_rank_ordering}
\title{Validate that rankings are consistent with ordering}
\usage{
validate_rank_ordering(rankings, linear_ordering)
}
\arguments{
\item{rankings}{A rank matrix of size n_assessors x n_items.}

\item{linear_ordering}{A list of length n_assessors. Each list element must
be a vector of orderings. It is assumed that the first vector element is
least favored, and the last element is most preferred.}
}
\value{
A logical. `TRUE` if consistent, `FALSE` otherwise.
}
\description{
We need a function to check that ranks are consistent with linear ordering We
assume the linear_ordering has n_assessors elements, each being a vector.
}
