% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assess_convergence.R
\name{assess_convergence}
\alias{assess_convergence}
\title{Create plots for assessing convergence of Metropolis-Hastings algorithm}
\usage{
assess_convergence(R, metric = "footrule", lambda = 0.1, nmc = 3000,
  burnin = 2000, L = 1, sd_alpha = 0.1, alpha_init = 1,
  max_items = 20, rho_displays = 4)
}
\arguments{
\item{R}{Ranks}

\item{metric}{string}

\item{lambda}{Prior}

\item{nmc}{Number of Monte Carlo samples in test run}

\item{burnin}{How many should we regard as burnin?}

\item{L}{Leap-and-shift step size}

\item{sd_alpha}{Standard deviation of alpha proposal}

\item{alpha_init}{Initial value of alpha}

\item{max_items}{The maximum number of items to plot in the diagnostic for rho.}

\item{rho_displays}{The maximum number of displays in the diagnostic plots for rho.}
}
\value{
A list containing two plots, one for alpha and one for rho.
}
\description{
Create plots for assessing convergence of Metropolis-Hastings algorithm
}
\examples{
# Let us first look at the convergence of alpha, for the potato data
# First we try with an alpha standard deviation of 5
check <- assess_convergence(R = potato_weighing, sd_alpha = 5)
# Now show the convergence plot for alpha
check$alpha_plot
# The acceptance rate was a bit low, so let us try to decrease the sd_alpha
check <- assess_convergence(R = potato_weighing, sd_alpha = 0.1)
# Then view the plot
check$alpha_plot
# This looked better!

# Next, we assess the convergence for rho.
# We start by using the results of the previous run.
check$rho_plot
# Let us see how increasing L changes the convergence
check <- assess_convergence(R = potato_weighing, sd_alpha = 0.1, L = 3)
# The acceptance rate went down. We should stick to L = 1.
check$rho_plot
}
