% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assess_convergence.R
\name{assess_convergence}
\alias{assess_convergence}
\title{Assessing convergence of Metropolis-Hastings algorithm}
\usage{
assess_convergence(model_fit, type = "alpha", items = NULL, k = NULL,
  assessors = NULL)
}
\arguments{
\item{model_fit}{A fitted model object of class \code{BayesMallows}, obtained
with \code{\link{compute_mallows}}.}

\item{type}{Should we plot \code{"alpha"}, \code{"rho"}, or
\code{"augmentation"}.}

\item{items}{The items to study in the diagnostic plot for \code{rho}. Either
a vector of item names, corresponding to \code{model_fit$item_names} or a
vector of indices. If NULL, five items are selected randomly.}

\item{k}{Window size for rolling mean. Used when \code{type =
"augmentation"}.}

\item{assessors}{The assessors to study in the diagnostic plot for
\code{"augmentation"}.}
}
\description{
Assessing convergence of Metropolis-Hastings algorithm
}
\seealso{
\code{\link{compute_mallows}}, \code{\link{plot.BayesMallows}}
}
