% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{beach_preferences}
\alias{beach_preferences}
\title{Beach Preferences}
\format{
An object of class \code{data.frame} with 1442 rows and 3 columns.
}
\usage{
beach_preferences
}
\description{
Example dataset from \insertCite{vitelli2018}{BayesMallows}, Section 6.2.
}
\references{
\insertAllCited{}
}
\seealso{
Other datasets: 
\code{\link{potato_true_ranking}},
\code{\link{potato_visual}},
\code{\link{potato_weighing}},
\code{\link{sushi_rankings}}
}
\concept{datasets}
\keyword{datasets}
