% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bernoulli_data}
\alias{bernoulli_data}
\title{Simulated intransitive pairwise preferences}
\format{
An object of class \code{data.frame} with 2280 rows and 3 columns.
}
\usage{
bernoulli_data
}
\description{
Simulated dataset based on the \link{potato_visual} data. Based on
the rankings in \link{potato_visual}, all n-choose-2 = 190 pairs of items were
sampled from each assessor. With probability .9, the pairwise
preference was in agreement with \link{potato_visual}, and with probability .1,
they were in disagreement. Hence, the data generating mechanism was a
Bernoulli error model \insertCite{crispino2019}{BayesMallows} with
\eqn{\theta=0.1}.
}
\seealso{
Other datasets: 
\code{\link{beach_preferences}},
\code{\link{cluster_data}},
\code{\link{potato_true_ranking}},
\code{\link{potato_visual}},
\code{\link{potato_weighing}},
\code{\link{sounds}},
\code{\link{sushi_rankings}}
}
\concept{datasets}
\keyword{datasets}
