% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/burnin.R
\name{burnin}
\alias{burnin}
\alias{burnin.BayesMallows}
\alias{burnin.BayesMallowsMixtures}
\alias{burnin.SMCMallows}
\title{See the burnin}
\usage{
burnin(model, ...)

\method{burnin}{BayesMallows}(model)

\method{burnin}{BayesMallowsMixtures}(model)

\method{burnin}{SMCMallows}(model)
}
\arguments{
\item{model}{A model object.}

\item{...}{Optional arguments passed on to other methods. Currently not used.}
}
\value{
An integer specifying the burnin, if it exists. Otherwise \code{NULL}.
}
\description{
See the current burnin value of the model.
}
\examples{
/inst/examples/burnin_example.R
}
\seealso{
Other modeling: 
\code{\link{burnin<-}()},
\code{\link{compute_mallows}()},
\code{\link{compute_mallows_mixtures}()},
\code{\link{compute_mallows_sequentially}()},
\code{\link{sample_prior}()},
\code{\link{update_mallows}()}
}
\concept{modeling}
