% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cp_consensus.R
\name{compute_cp_consensus}
\alias{compute_cp_consensus}
\title{Compute CP Consensus}
\usage{
compute_cp_consensus(model_fit, burnin)
}
\arguments{
\item{model_fit}{An object returned from \code{\link{compute_mallows}}.}

\item{burnin}{The number of iterations to discard as burn-in.}
}
\description{
Compute CP Consensus
}
\details{
Computes the cumulative probability (CP) consensus rankings of the latent ranks.
If the number of clusters is larger than 1, it returns the CP consensus per cluster.
}
