% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_cp_consensus.R
\name{compute_cp_consensus}
\alias{compute_cp_consensus}
\title{Compute CP Consensus}
\usage{
compute_cp_consensus(model_fit, burnin)
}
\arguments{
\item{model_fit}{An object returned from \code{\link{compute_mallows}}.}

\item{burnin}{A numeric value specifying the number of iterations to discard
as burn-in. See \code{\link{assess_convergence}}.}
}
\description{
Compute the cumulative probability (CP) consensus ranking
\insertCite{vitelli2018}{BayesMallows} of the items. For mixture models, the
CP consensus is given for each ranking.
}
\references{
\insertAllCited{}
}
