% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{compute_importance_sampling_estimate}
\alias{compute_importance_sampling_estimate}
\title{Compute importance sampling estimates of log partition function
for footrule and Spearman distances.}
\usage{
compute_importance_sampling_estimate(alpha_vector, n,
  metric = "footrule", nmc = 1000000L)
}
\arguments{
\item{alpha_vector}{Vector of alpha values at which to compute partition function.}

\item{n}{Integer specifying the number of ranked items.}

\item{metric}{Distance measure of the target Mallows distribution. Defaults to \code{footrule}.}

\item{nmc}{Number of Monte Carlo samples. Defaults to \code{1e6}.}
}
\description{
Compute importance sampling estimates of log partition function
for footrule and Spearman distances.
}
