% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_mallows.R
\name{compute_mallows}
\alias{compute_mallows}
\title{Compute the posterior distribution for the Mallows model.}
\usage{
compute_mallows(R, metric = "footrule", lambda = 0.1, nmc = 3000,
  burnin = 2000, L = NULL, sd_alpha = 0.1, alpha_init = 5)
}
\arguments{
\item{R}{A matrix of ranked items.}

\item{metric}{The distance metric to use. Currently available ones are
\code{"footrule"}, \code{"spearman"}, \code{"kendall"}, \code{"cayley"},
and \code{"hamming"}.}

\item{lambda}{Parameter for the prior distribution of \code{alpha}. Defaults to 0.1.}

\item{nmc}{Number of Monte Carlo samples to keep.}

\item{burnin}{Number of samples to discard as burn-in before collecting \code{nmc} samples.}

\item{L}{Step size of the leap-and-shift proposal distribution. Defaults to NULL, which
means that it is automatically set to n/5.}

\item{sd_alpha}{Standard deviation of the proposal distribution for alpha.}

\item{alpha_init}{Initial value of alpha.}
}
\value{
A list of class BayesMallows.
}
\description{
Compute the posterior distribution for the Mallows model.
}
\details{
It is usually a
good idea to first use \code{\link{assess_convergence}} to determine the
algorithm parameters.
}
\examples{
# Compute the posterior
model_fit <- compute_mallows(potato_weighing, "footrule", nmc = 10000, burnin = 5000)
# Plot the posterior histogram for alpha
plot(model_fit, type = "alpha", bins = 50)
}
\seealso{
\code{\link{assess_convergence}}, \code{\link{plot.BayesMallows}}.
}
