% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_mallows.R
\name{compute_mallows}
\alias{compute_mallows}
\title{Compute the posterior distribution for the Mallows model.}
\usage{
compute_mallows(R = NULL, P = NULL, metric = "footrule",
  lambda = 0.1, nmc = 3000, L = NULL, sd_alpha = 0.1,
  alpha_init = 1, alpha_jump = 1, thinning = 1,
  aug_diag_thinning = 100)
}
\arguments{
\item{R}{A matrix of ranked items. See \code{\link{create_ranking}} if you
have an ordered set of items that needs to be converted to rankings. If
\code{P} is provided, \code{R} is an optional initial value of the rankings,
generated by \code{\link{generate_initial_ranking}}.}

\item{P}{A dataframe with the transitive closure of pairwise comparisons. If
it has class \code{BayesMallowsTC}, generated by
\code{\link{generate_transitive_closure}}, this will be used. Otherwise,
\code{\link{generate_transitive_closure}} will be called on this dataframe.}

\item{metric}{The distance metric to use. Currently available ones are
\code{"footrule"}, \code{"spearman"}, \code{"kendall"}, \code{"cayley"},
and \code{"hamming"}.}

\item{lambda}{Parameter for the prior distribution of \code{alpha}. Defaults
to 0.1.}

\item{nmc}{Number of Monte Carlo samples to keep.}

\item{L}{Step size of the leap-and-shift proposal distribution. Defaults to
NULL, which means that it is automatically set to n/5.}

\item{sd_alpha}{Standard deviation of the proposal distribution for alpha.}

\item{alpha_init}{Initial value of alpha.}

\item{alpha_jump}{How many times should we sample \code{rho} between each
time we sample \code{alpha}. Setting \code{alpha_jump} to a high number can
significantly speed up computation time, since we then do not have to do
expensive computation of the partition function.}

\item{thinning}{Keep every \code{thinning} iteration of \code{rho}.}

\item{aug_diag_thinning}{The interval in which we save augmentation
diagnostics.}
}
\value{
A list of class BayesMallows.
}
\description{
Compute the posterior distribution for the Mallows model.
}
\examples{
# Compute the posterior
model_fit <- compute_mallows(potato_weighing, "footrule", nmc = 1000)

}
\references{
\insertRef{vitelli2018}{BayesMallows}
}
\seealso{
\code{\link{assess_convergence}}, \code{\link{plot.BayesMallows}}.
}
