% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_mallows_sequentially.R
\name{compute_mallows_sequentially}
\alias{compute_mallows_sequentially}
\title{Estimate the Bayesian Mallows Model Sequentially}
\usage{
compute_mallows_sequentially(
  data,
  initial_values,
  model_options = set_model_options(),
  smc_options = set_smc_options(),
  compute_options = set_compute_options(),
  priors = set_priors(),
  pfun_estimate = NULL
)
}
\arguments{
\item{data}{A list of objects of class "BayesMallowsData" returned from
\code{\link[=setup_rank_data]{setup_rank_data()}}. Each list element is interpreted as the data belonging
to a given timepoint.}

\item{initial_values}{An object of class "BayesMallowsPriorSamples" returned
from \code{\link[=sample_prior]{sample_prior()}}.}

\item{model_options}{An object of class "BayesMallowsModelOptions" returned
from \code{\link[=set_model_options]{set_model_options()}}.}

\item{smc_options}{An object of class "SMCOptions" returned from
\code{\link[=set_smc_options]{set_smc_options()}}.}

\item{compute_options}{An object of class "BayesMallowsComputeOptions"
returned from \code{\link[=set_compute_options]{set_compute_options()}}.}

\item{priors}{An object of class "BayesMallowsPriors" returned from
\code{\link[=set_priors]{set_priors()}}.}

\item{pfun_estimate}{Object returned from \code{\link[=estimate_partition_function]{estimate_partition_function()}}.
Defaults to \code{NULL}, and will only be used for footrule, Spearman, or
Ulam distances when the cardinalities are not available, cf.
\code{\link[=get_cardinalities]{get_cardinalities()}}.}
}
\value{
An object of class SMCMallows.
}
\description{
Compute the posterior distributions of the parameters of the
Bayesian Mallows model using sequential Monte Carlo. This is based on the
algorithms developed in
\insertCite{steinSequentialInferenceMallows2023;textual}{BayesMallows}.
}
\examples{
# Observe one ranking at each of 12 timepoints
data <- lapply(seq_len(nrow(potato_visual)), function(i) {
  setup_rank_data(potato_visual[i, ])
})

initial_values <- sample_prior(200, 20)
smc_options <- set_smc_options(n_particles = 4)
}
\references{
\insertAllCited{}
}
\seealso{
Other modeling: 
\code{\link{burnin}()},
\code{\link{burnin<-}()},
\code{\link{compute_mallows}()},
\code{\link{compute_mallows_mixtures}()},
\code{\link{sample_prior}()},
\code{\link{update_mallows}()}
}
\concept{modeling}
