% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_map_consensus.R
\name{compute_map_consensus}
\alias{compute_map_consensus}
\title{Maximum a Posterior Consensus Ranking}
\usage{
compute_map_consensus(model_fit, burnin = model_fit$burnin)
}
\arguments{
\item{model_fit}{An object of type \code{BayesMallows}, returned from
\code{\link{compute_mallows}}.}

\item{burnin}{A numeric value specifying the number of iterations
to discard as burn-in. Defaults to \code{model_fit$burnin}, and must be
provided if \code{model_fit$burnin} does not exist. See \code{\link{assess_convergence}}.}
}
\description{
Compute the maximum a posterior consensus ranking. If a mixture model
is supplied, the maximum a posterior consensus ranking for each mixture
is returned.
}
