% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_posterior.R
\name{compute_posterior}
\alias{compute_posterior}
\title{Compute the posterior distribution for the Mallows model.}
\usage{
compute_posterior(R, metric = "footrule", lambda = 0.1, nmc = 1000,
  L = 1, sd_alpha = 0.1, alpha_init = 5)
}
\arguments{
\item{R}{A matrix of ranks, N x n. This matrix is transposed before calling
the underlying C++ function.}

\item{metric}{The metric to use.}
}
\value{
TBD
}
\description{
Compute the posterior distribution for the Mallows model.
}
\examples{
model_fit <- compute_posterior(potato_weighing, "footrule")
}
