% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_posterior_intervals.R
\name{compute_posterior_intervals}
\alias{compute_posterior_intervals}
\title{Compute Posterior Intervals}
\usage{
compute_posterior_intervals(model_fit, burnin, parameter, level = 0.95,
  decimals = 3)
}
\arguments{
\item{model_fit}{An object returned from \code{\link{compute_mallows}}.}

\item{burnin}{The number of iterations to discard as burn-in.}

\item{parameter}{Which parameter to compute posterior intervals for. One of
\code{"alpha"}, \code{"rho"}, or \code{"cluster_probs"}.}

\item{level}{Confidence level.}

\item{decimals}{Number of decimals to include in posterior intervals and the mean
and median.}
}
\description{
Compute Posterior Intervals
}
\details{
This function computes both the Highest Posterior Density Interval (HPDI),
which may be discontinuous for bimodal distributions, and
the central posterior interval, which is simply defined by the quantiles of the posterior
distribution.
}
