% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_processing_functions.R
\name{compute_posterior_intervals_alpha}
\alias{compute_posterior_intervals_alpha}
\title{Compute Posterior Intervals Alpha}
\usage{
compute_posterior_intervals_alpha(output, nmc, burnin)
}
\arguments{
\item{output}{output}

\item{nmc}{nmc}

\item{burnin}{burn-in}
}
\description{
posterior confidence intervals
}
