% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smc_post_processing_functions.R
\name{compute_posterior_intervals_alpha}
\alias{compute_posterior_intervals_alpha}
\title{Compute Posterior Intervals Alpha}
\usage{
compute_posterior_intervals_alpha(output, nmc, burnin, verbose = FALSE)
}
\arguments{
\item{output}{a subset of an SMCMallows object (though technically any matrix
will do)}

\item{nmc}{Number of Monte Carlo samples}

\item{burnin}{A numeric value specifying the number of iterations
to discard as burn-in.}

\item{verbose}{if \code{TRUE}, prints the final output even if the function
is assigned to an object. Defaults to \code{FALSE}.}
}
\description{
posterior confidence intervals
}
\seealso{
Other deprecated: 
\code{\link{compute_posterior_intervals_rho}()},
\code{\link{compute_rho_consensus}()}
}
\concept{deprecated}
