% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smc_post_processing_functions.R
\name{compute_posterior_intervals_rho}
\alias{compute_posterior_intervals_rho}
\title{Compute Posterior Intervals Rho}
\usage{
compute_posterior_intervals_rho(output, nmc, burnin, verbose = FALSE)
}
\arguments{
\item{output}{output}

\item{nmc}{nmc}

\item{burnin}{burn-in}

\item{verbose}{if \code{TRUE}, prints the final output even if the function is assigned to an object. Defaults to \code{FALSE}.}
}
\description{
posterior confidence intervals for rho
}
\author{
Anja Stein
}
