% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smc_post_processing_functions.R
\name{compute_rho_consensus}
\alias{compute_rho_consensus}
\title{Compute rho consensus}
\usage{
compute_rho_consensus(
  output,
  nmc,
  burnin,
  C,
  type = "CP",
  colnames = NULL,
  verbose = FALSE
)
}
\arguments{
\item{output}{Matrix}

\item{nmc}{Number of Monte Carlo samples}

\item{burnin}{Burnin}

\item{C}{C Number of clusters.}

\item{type}{type}

\item{colnames}{Column names}

\item{verbose}{Logical}
}
\description{
This function has been deprecated. Please use
  \code{\link{compute_consensus.SMCMallows}}.

MAP and CP consensus ranking estimates

  This function is deprecated. Please use
  \code{\link{compute_consensus.SMCMallows}} instead.
}
\seealso{
Other deprecated: 
\code{\link{compute_posterior_intervals_alpha}()},
\code{\link{compute_posterior_intervals_rho}()}
}
\concept{deprecated}
