% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smc_post_processing_functions.R
\name{compute_rho_consensus}
\alias{compute_rho_consensus}
\title{Compute rho consensus}
\usage{
compute_rho_consensus(
  output,
  nmc,
  burnin,
  C,
  type = "CP",
  colnames = NULL,
  verbose = FALSE
)
}
\arguments{
\item{output}{a subset of an SMCMallows object (though technically any matrix will do)}

\item{nmc}{Number of Monte Carlo samples}

\item{burnin}{A numeric value specifying the number of iterations
to discard as burn-in.}

\item{C}{C}

\item{type}{type}

\item{colnames}{colnames}

\item{verbose}{if \code{TRUE}, prints the final output even if the function
is assigned to an object. Defaults to \code{FALSE}.}
}
\description{
MAP AND CP consensus ranking estimates
}
\seealso{
\code{\link{smc_mallows_new_item_rank}} and
\code{\link{smc_mallows_new_users}}, which are functions generating objects
of SMCMallows class.
}
\author{
Anja Stein
}
