% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smc_post_processing_functions.R
\name{compute_rho_consensus}
\alias{compute_rho_consensus}
\title{Compute rho consensus}
\usage{
compute_rho_consensus(output, nmc, burnin, C, type, verbose = FALSE)
}
\arguments{
\item{output}{output}

\item{nmc}{nmc}

\item{burnin}{burn-in}

\item{C}{C}

\item{type}{type}

\item{verbose}{if \code{TRUE}, prints the final output even if the function is assigned to an object. Defaults to \code{FALSE}.}
}
\description{
MAP AND CP consensus ranking estimates
}
