% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_processing_functions.R
\name{compute_rho_consensus}
\alias{compute_rho_consensus}
\title{Compute rho consensus}
\usage{
compute_rho_consensus(output, nmc, burnin, C, type)
}
\arguments{
\item{output}{output}

\item{nmc}{nmc}

\item{burnin}{burn-in}

\item{C}{C}

\item{type}{type}
}
\description{
MAP AND CP consensus ranking estimates
}
