% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rank_conversion.R
\name{create_ranking}
\alias{create_ranking}
\title{Convert between ranking and ordering.}
\usage{
create_ranking(O)
}
\arguments{
\item{O}{A vector or matrix of ordered items. If a matrix,
it should be of size N times n, where N is the number of samples
and n is the number of items.}
}
\value{
A vector or matrix of rankings.
}
\description{
Convert between ranking and ordering.
}
\examples{
# A vector of ordered items.
O <- c(5, 1, 2, 4, 3)
# Get ranks
R <- create_ranking(O)
# R is c(2, 3, 5, 4, 1)

# Next, we have a matrix with N = 19 samples
# and n = 4 items
set.seed(21)
N <- 10
n <- 4
O <- t(replicate(N, sample.int(n)))
# Convert the ordering to ranking
R <- create_ranking(O)

}
