% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_partition_function.R
\name{estimate_partition_function}
\alias{estimate_partition_function}
\title{Estimate Partition Function with Importance Sampling}
\usage{
estimate_partition_function(alpha_vector, n_items, metric, nmc, degree)
}
\arguments{
\item{alpha_vector}{Numeric vector of \eqn{\alpha} values over which
to compute the importance sampling estimate.}

\item{n_items}{Integer specifying the number of items.}

\item{metric}{Character string specifying the distance measure to use.
Available options are \code{"footrule"}, \code{"spearman"},
\code{"kendall"}, \code{"cayley"}, and \code{"hamming"}.}

\item{nmc}{Integer specifying the number of Monte Carlo samples to use in the
importance sampling.}

\item{degree}{Integer specifying the degree of the polynomial used to estimate
\eqn{\log(\alpha)} from the grid of values provided by the importance sampling
estimate.}
}
\value{
A vector of length \code{degree} which can be supplied to the
\code{is_fit} argument of \code{\link{compute_mallows}}.
}
\description{
\code{estimate_partition_function} uses importance sampling to estimate
the logarithm of the partition function of the Mallows rank model.
}
