% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_initial_ranking.R
\name{generate_initial_ranking}
\alias{generate_initial_ranking}
\title{Generic function for generating initial ranking}
\usage{
generate_initial_ranking(preferences, n_items, cl)
}
\description{
Generic function for generating initial ranking
}
