% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_initial_ranking.R
\name{generate_initial_ranking}
\alias{generate_initial_ranking}
\title{Title}
\usage{
generate_initial_ranking(tc, n_items = max(tc[, c("bottom_item",
  "top_item")]))
}
\arguments{
\item{tc}{A dataframe with the transitive closure of pairwise comparisons. If
it has class \code{BayesMallowsTC}, generated by
\code{\link{generate_transitive_closure}}, this will be used. Otherwise,
\code{\link{generate_transitive_closure}} will be called on this dataframe.}

\item{n_items}{The total number of items. If not provided, it is assumed to
equal to \code{max(tc[, c("bottom_item", "top_item")])}.}
}
\description{
Title
}
