% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_transitive_closure.R
\name{generate_transitive_closure}
\alias{generate_transitive_closure}
\title{Generate transitive closure}
\usage{
generate_transitive_closure(df)
}
\arguments{
\item{df}{Data frame with columns \code{assessor}, \code{top_item}, and \code{bottom_item}.}
}
\description{
Generate transitive closure
}
