% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_transitive_closure.R
\name{generate_transitive_closure}
\alias{generate_transitive_closure}
\title{Title}
\usage{
generate_transitive_closure(df)
}
\arguments{
\item{df}{Dataframe with columns \code{assessor}, \code{bottom_item}, and
\code{top_item}.}
}
\description{
Title
}
