% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{get_summation_distances}
\alias{get_summation_distances}
\title{Get the distances for computing the partition function given
the cardinalities.}
\usage{
get_summation_distances(n, cardinalities, metric = "footrule")
}
\arguments{
\item{n}{The number of items.}

\item{cardinalities}{Number of times each distance appears.}

\item{metric}{A string. Available options are \code{"footrule"},
\code{"kendall"}, and \code{"spearman"}. Defaults to \code{"footrule"}.}
}
\value{
A scalar, the logarithm of the partition function.
}
\description{
Get the distances for computing the partition function given
the cardinalities.
}
\keyword{internal}
