% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heat_plot.R
\name{heat_plot}
\alias{heat_plot}
\title{Heat plot of posterior probabilities}
\usage{
heat_plot(model_fit, burnin = model_fit$burnin, ...)
}
\arguments{
\item{model_fit}{An object of type \code{BayesMallows}, returned from
\code{\link[=compute_mallows]{compute_mallows()}}.}

\item{burnin}{A numeric value specifying the number of iterations
to discard as burn-in. Defaults to \code{x$burnin}, and must be
provided if \code{x$burnin} does not exist. See \code{\link[=assess_convergence]{assess_convergence()}}.}

\item{...}{Additional arguments passed on to other methods. In particular,
\code{type = "CP"} or \code{type = "MAP"} can be passed on to
\code{\link[=compute_consensus]{compute_consensus()}} to determine the order of items along the
horizontal axis.}
}
\value{
A ggplot object.
}
\description{
Generates a heat plot with items in their consensus ordering along the
horizontal axis and ranking along the vertical axis. The color denotes
posterior probability.
}
\examples{
set.seed(1)
model_fit <- compute_mallows(
  setup_rank_data(potato_visual),
  compute_options = set_compute_options(nmc = 2000, burnin = 500))

heat_plot(model_fit)
heat_plot(model_fit, type = "MAP")
}
\seealso{
Other posterior quantities: 
\code{\link{assign_cluster}()},
\code{\link{compute_consensus}()},
\code{\link{compute_posterior_intervals}()},
\code{\link{plot.BayesMallows}()},
\code{\link{plot.SMCMallows}()},
\code{\link{plot_elbow}()},
\code{\link{plot_top_k}()},
\code{\link{predict_top_k}()},
\code{\link{print.BayesMallows}()}
}
\concept{posterior quantities}
