% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smc_metropolis_hastings_alpha.R
\name{metropolis_hastings_alpha}
\alias{metropolis_hastings_alpha}
\title{Metropolis-Hastings Alpha}
\usage{
metropolis_hastings_alpha(alpha, n_items, rankings, metric, rho, logz_estimate)
}
\arguments{
\item{alpha}{Numeric value og the scale parameter}

\item{n_items}{Integer is the number of items in a ranking}

\item{rankings}{A matrix of size \eqn{N }\eqn{\times}{x}\eqn{ n_items} of
rankings in each row. Alternatively, if \eqn{N} equals 1, \code{rankings}
can be a vector.}

\item{metric}{A character string specifying the distance metric to use in the
Bayesian Mallows Model. Available options are \code{"footrule"},
\code{"spearman"}, \code{"cayley"}, \code{"hamming"}, \code{"kendall"}, and
\code{"ulam"}.}

\item{rho}{Numeric vector specifying the consensus ranking}

\item{logz_estimate}{Estimate of the partition function, computed with
\code{\link{estimate_partition_function}} in the BayesMallow R package {estimate_partition_function}.}
}
\value{
\code{alpha} or \code{alpha_prime}: Numeric value to be used as the proposal of a new alpha
}
\description{
Function to perform Metropolis-Hastings for new alpha under the Mallows model
}
