% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smc_metropolis_hastings_rho.R
\name{metropolis_hastings_rho}
\alias{metropolis_hastings_rho}
\title{Metropolis-Hastings Rho}
\usage{
metropolis_hastings_rho(alpha, n_items, rankings, metric, rho, leap_size)
}
\arguments{
\item{alpha}{Numeric value og the scale parameter}

\item{n_items}{Integer is the number of items in a ranking}

\item{rankings}{A matrix of size \eqn{N }\eqn{\times}{x}\eqn{ n_items} of
rankings in each row. Alternatively, if \eqn{N} equals 1, \code{rankings}
can be a vector.}

\item{metric}{# TODO: describe}

\item{rho}{A ranking sequence vector}

\item{leap_size}{Integer specifying the step size of the leap-and-shift
proposal distribution.}
}
\value{
\code{rho} or \code{rho_prime}: A ranking sequence vector to be the next value of rho in the MCMC chain
}
\description{
Function to perform Metropolis-Hastings for new rho under the Mallows model with footrule distance metric!
}
\examples{
rho <- c(1,2,3,4,5,6)
alpha <- 2
metric <- "footrule"
n_items <- 6

metropolis_hastings_rho(
	alpha = alpha, n_items = n_items, rankings = rho, metric = metric,
	rho = rho, leap_size = 1
)

metropolis_hastings_rho(
	alpha = alpha, n_items = n_items, rankings = rho, metric = metric,
	rho = rho, leap_size = 2
)

metropolis_hastings_rho(
	alpha = alpha, n_items = n_items, rankings = rho, metric = metric,
	rho = rho, leap_size = 3
)

rankings <- sample_mallows(
 rho0 = rho, alpha0 = alpha, n_samples = 10, burnin = 1000, thinning = 500
)
metropolis_hastings_rho(
	alpha = alpha, n_items = n_items, rankings = rankings, metric = metric,
	rho = rho, leap_size = 1
)

}
\author{
Anja Stein
}
