% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.BayesMallows.R
\name{plot.BayesMallows}
\alias{plot.BayesMallows}
\title{Posterior plots}
\usage{
\method{plot}{BayesMallows}(fit, type = "alpha", items = NULL, ...)
}
\arguments{
\item{fit}{An object of type \code{BayesMallows}, returned from
\code{\link{compute_posterior}}.}

\item{type}{Which type of plot. Either \code{"alpha"} for a posterior histogram
of the scale parameter or \code{"rho"} for a posterior histogram of a ranked
item.}

\item{items}{If \code{type = "rho"} this argument must be set. It defines
which item(s) to plot. If \code{type = "alpha"} this argument is ignored.}

\item{...}{Other arguments passed to ggplot2::geom_histogram.}
}
\value{
A plot.
}
\description{
Posterior plots
}
\examples{
# Fit a model to the potato data
model_fit <- compute_posterior(potato_weighing, "footrule",
nmc = 10000, burnin = 5000)
# Plot the posterior histogram of alpha
plot(model_fit, type = "alpha", bins = 50)
# We can also plot the posterior histograms of rankings of items
plot(model_fit, type = "rho", items = c(2, 5), bins = 20)

}
