% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.BayesMallows.R
\name{plot.BayesMallows}
\alias{plot.BayesMallows}
\title{Plot Posterior Distributions}
\usage{
\method{plot}{BayesMallows}(x, burnin, type = "alpha", items = NULL,
  ...)
}
\arguments{
\item{x}{An object of type \code{BayesMallows}, returned from
\code{\link{compute_mallows}}.}

\item{burnin}{A numeric value specifying the number of iterations
to discard as burn-in. See \code{\link{assess_convergence}}.}

\item{type}{Character string defining the parameter to plot. Available
options are \code{"alpha"}, \code{"rho"}, \code{"cluster_probs"}, and
\code{"cluster_assignment"}.}

\item{items}{The items to study in the diagnostic plot for \code{rho}. Either
a vector of item names, corresponding to \code{x$items} or a
vector of indices. If NULL, five items are selected randomly.
Only used when \code{type = "rho"}.}

\item{...}{Other arguments passed to \code{plot} (not used).}
}
\description{
Plot posterior distributions of the parameters of the Mallows Rank model.
}
