% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.BayesMallows.R
\name{plot.SMCMallows}
\alias{plot.SMCMallows}
\title{Plot SMC Posterior Distributions}
\usage{
\method{plot}{SMCMallows}(x, parameter = "alpha", items = NULL, ...)
}
\arguments{
<<<<<<< HEAD
\item{x}{An object of type \code{SMC-Mallows}, returned for example from
\code{\link[=smc_mallows_new_users]{smc_mallows_new_users()}}.}

\item{nmc}{Number of Monte Carlo samples}

\item{burnin}{A numeric value specifying the number of iterations
to discard as burn-in. Defaults to \code{model_fit$burnin}, and must be
provided if \code{model_fit$burnin} does not exist. See
\code{\link[=assess_convergence]{assess_convergence()}}.}
=======
\item{x}{An object of type \code{SMC-Mallows}.}
>>>>>>> smc-issue-306

\item{parameter}{Character string defining the parameter to plot. Available
options are \code{"alpha"} and \code{"rho"}.}

\item{items}{Either a vector of item names, or a vector of indices. If NULL,
five items are selected randomly.}

\item{...}{Other arguments passed to \code{\link[base:plot]{base::plot()}} (not used).}
}
\value{
A plot of the posterior distributions
}
\description{
Plot posterior distributions of SMC-Mallow parameters.
}
\seealso{
Other posterior quantities: 
\code{\link{assign_cluster}()},
\code{\link{compute_consensus.BayesMallows}()},
\code{\link{compute_consensus}()},
\code{\link{compute_posterior_intervals.BayesMallows}()},
\code{\link{compute_posterior_intervals}()},
\code{\link{heat_plot}()},
\code{\link{plot.BayesMallows}()},
\code{\link{plot_elbow}()},
\code{\link{plot_top_k}()},
\code{\link{predict_top_k}()},
\code{\link{print.BayesMallowsMixtures}()},
\code{\link{print.BayesMallows}()}
}
\concept{posterior quantities}
