% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.SMCMallows.R
\name{plot.SMCMallows}
\alias{plot.SMCMallows}
\title{Plot SMC Posterior Distributions}
\usage{
\method{plot}{SMCMallows}(
  x,
  nmc = nrow(x$rho_samples[, 1, ]),
  burnin = 0,
  parameter = "alpha",
  time = ncol(x$rho_samples[, 1, ]),
  C = 1,
  colnames = NULL,
  items = NULL,
  ...
)
}
\arguments{
\item{x}{An object of type \code{SMC-Mallows}, returned for example from
\code{\link{smc_mallows_new_users}}.}

\item{nmc}{Number of Monte Carlo samples}

\item{burnin}{A numeric value specifying the number of iterations
to discard as burn-in. Defaults to \code{model_fit$burnin}, and must be
provided if \code{model_fit$burnin} does not exist. See
\code{\link{assess_convergence}}.}

\item{parameter}{Character string defining the parameter to plot. Available
options are \code{"alpha"} and \code{"rho"}.}

\item{time}{Integer determining the update slice to plot}

\item{C}{Number of cluster}

\item{colnames}{A vector of item names. If NULL, generic names are generated
for the items in the ranking.}

\item{items}{Either a vector of item names, or a vector of indices. If NULL,
five items are selected randomly.}

\item{...}{Other arguments passed to \code{\link[base]{plot}} (not used).}
}
\value{
A plot of the posterior distributions
}
\description{
Plot posterior distributions of SMC-Mallow parameters.
}
\examples{
set.seed(994)

n_items <- dim(sushi_rankings)[2]
metric <- "footrule"

# Perform the resample-move SMC algorithm
smc_test <- smc_mallows_new_users(
  R_obs = sushi_rankings[1:100, ], type = "complete", n_items = n_items,
  metric = metric, leap_size = floor(n_items / 5), N = 100, Time = 10,
  mcmc_kernel_app = 5, num_new_obs = 5,
  alpha_prop_sd = 0.5, lambda = 0.15, alpha_max = 1e3
)

# Plot rho
plot(smc_test, colnames = colnames(sushi_rankings), parameter = "rho")

# Plot alpha
plot(smc_test, parameter = "alpha")
}
\seealso{
Other posterior quantities: 
\code{\link{assign_cluster}()},
\code{\link{compute_consensus.BayesMallows}()},
\code{\link{compute_consensus.SMCMallows}()},
\code{\link{compute_consensus}()},
\code{\link{compute_posterior_intervals.BayesMallows}()},
\code{\link{compute_posterior_intervals.SMCMallows}()},
\code{\link{compute_posterior_intervals}()},
\code{\link{heat_plot}()},
\code{\link{plot.BayesMallows}()},
\code{\link{plot_elbow}()},
\code{\link{plot_top_k}()},
\code{\link{predict_top_k}()},
\code{\link{print.BayesMallowsMixtures}()},
\code{\link{print.BayesMallows}()}
}
\author{
Waldir Leoncio
}
\concept{posterior quantities}
