% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smc_post_processing_functions.R
\name{plot_alpha_posterior}
\alias{plot_alpha_posterior}
\title{Plot Alpha Posterior}
\usage{
plot_alpha_posterior(output, nmc, burnin, verbose = FALSE)
}
\arguments{
\item{output}{output}

\item{nmc}{nmc}

\item{burnin}{burn-in}

\item{verbose}{if \code{TRUE}, prints the final output even if the function is assigned to an object. Defaults to \code{FALSE}.}
}
\description{
posterior for alpha
}
\author{
Anja Stein
}
