% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.SMCMallows.R
\name{plot_alpha_posterior}
\alias{plot_alpha_posterior}
\title{Plot Alpha Posterior}
\usage{
plot_alpha_posterior(output, nmc, burnin)
}
\arguments{
\item{output}{input}

\item{nmc}{Number of Monte Carlo samples}

\item{burnin}{A numeric value specifying the number of iterations
to discard as burn-in. Defaults to \code{model_fit$burnin}, and must be
provided if \code{model_fit$burnin} does not exist. See \code{\link{assess_convergence}}.}
}
\description{
posterior for alpha
}
\author{
Anja Stein
}
