% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_elbow.R
\name{plot_elbow}
\alias{plot_elbow}
\title{Posterior plots}
\usage{
plot_elbow(..., burnin)
}
\arguments{
\item{...}{One or more objects returned from \code{\link{compute_mallows}},
separated by comma, or a list of such objects.}

\item{burnin}{Number of iterations to discard as burn-in.}
}
\value{
A plot.
}
\description{
Posterior plots
}
