% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{potato_true_ranking}
\alias{potato_true_ranking}
\title{True ranking of the weights of 20 potatoes.}
\format{
An object of class \code{numeric} of length 20.
}
\usage{
potato_true_ranking
}
\description{
True ranking of the weights of 20 potatoes.
}
\references{
\insertRef{liu2019}{BayesMallows}
}
\seealso{
Other datasets: 
\code{\link{beach_preferences}},
\code{\link{potato_partial}},
\code{\link{potato_visual}},
\code{\link{potato_weighing}},
\code{\link{sample_dataset}},
\code{\link{sushi_rankings}}
}
\concept{datasets}
\keyword{datasets}
