% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{potato_true_ranking}
\alias{potato_true_ranking}
\title{True ranking of the weights of 20 potatoes.}
\format{
An object of class \code{numeric} of length 20.
}
\usage{
potato_true_ranking
}
\description{
True ranking of the weights of 20 potatoes.
}
\references{
\insertRef{liu2019}{BayesMallows}
}
\keyword{datasets}
