% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{potato_weighing}
\alias{potato_weighing}
\title{Result of ranking potatoes by weight, where the assessors were
allowed to lift the potatoes. 12 assessors ranked 20 potatoes.}
\format{
An object of class \code{matrix} (inherits from \code{array}) with 12 rows and 20 columns.
}
\usage{
potato_weighing
}
\description{
Result of ranking potatoes by weight, where the assessors were
allowed to lift the potatoes. 12 assessors ranked 20 potatoes.
}
\references{
\insertRef{liu2019}{BayesMallows}
}
\keyword{datasets}
