% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.BayesMallows.R
\name{print.BayesMallows}
\alias{print.BayesMallows}
\title{Print Method for BayesMallows Objects}
\usage{
\method{print}{BayesMallows}(x, ...)
}
\arguments{
\item{x}{An object of type \code{BayesMallows}, returned from
\code{\link{compute_mallows}}.}

\item{...}{Other arguments passed to \code{print} (not used).}
}
\description{
The default print method for a \code{BayesMallows} object.
}
