% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.BayesMallowsMixtures.R
\name{print.BayesMallowsMixtures}
\alias{print.BayesMallowsMixtures}
\title{Print Method for BayesMallowsMixtures Objects}
\usage{
\method{print}{BayesMallowsMixtures}(x, ...)
}
\arguments{
\item{x}{An object of type \code{BayesMallowsMixtures}, returned from
\code{\link{compute_mallows_mixtures}}.}

\item{...}{Other arguments passed to \code{print} (not used).}
}
\description{
The default print method for a \code{BayesMallowsMixtures} object.
}
