% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{run_mcmc}
\alias{run_mcmc}
\title{Worker function for computing the posterior distribtuion.}
\usage{
run_mcmc(rankings, nmc, preferences, constrained, cardinalities, is_fit,
  rho_init, metric = "footrule", n_clusters = 1L,
  include_wcd = FALSE, leap_size = 1L, alpha_prop_sd = 0.5,
  alpha_init = 5, alpha_jump = 1L, lambda = 0.1, psi = 10L,
  thinning = 1L, aug_diag_thinning = 100L,
  save_augmented_data = FALSE)
}
\arguments{
\item{rankings}{A set of complete rankings, with one sample per column.
With n_assessors samples and n_items items, rankings is n_items x n_assessors.}

\item{nmc}{Number of Monte Carlo samples.}

\item{preferences}{Matrix of preferences preferences, 3 x n_items.}

\item{constrained}{Matrix of constrained elements, 2 rows.}

\item{cardinalities}{Used when metric equals \code{"footrule"} or
\code{"spearman"} for computing the partition function. Defaults to
\code{R_NilValue}.}

\item{is_fit}{Importance sampling fit.}

\item{metric}{The distance metric to use. One of \code{"spearman"},
\code{"footrule"}, \code{"kendall"}, \code{"cayley"}, or
\code{"hamming"}.}

\item{n_clusters}{Number of clusters. Defaults to 1.}

\item{include_wcd}{Boolean defining whether or
not to store the within-cluster distance.}

\item{leap_size}{Leap-and-shift step size.}

\item{alpha_prop_sd}{Standard deviation of proposal distribution for alpha.}

\item{alpha_init}{Initial value of alpha.}

\item{alpha_jump}{How many times should we sample \code{rho} between
each time we sample \code{alpha}. Setting \code{alpha_jump} to a high
number can significantly speed up computation time, since we then do not
have to do expensive computation of the partition function.}

\item{lambda}{Parameter of the prior distribution.}

\item{psi}{Hyperparameter for the Dirichlet prior distribution used in clustering.}

\item{thinning}{Thinning parameter. Keep only every \code{thinning} rank
sample from the posterior distribution.}

\item{aug_diag_thinning}{The interval in which we save
augmentation diagnostics.}

\item{save_augmented_data}{Whether or not to save the augmented data every
\code{thinning}th iteration.}
}
\description{
Worker function for computing the posterior distribtuion.
}
\keyword{internal}
