% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{run_mcmc}
\alias{run_mcmc}
\title{Worker function for computing the posterior distribution.}
\usage{
run_mcmc(
  rankings,
  obs_freq,
  constraints,
  compute_options,
  priors,
  cardinalities,
  logz_estimate,
  rho_init,
  metric = "footrule",
  error_model = "none",
  n_clusters = 1L,
  alpha_init = 5,
  verbose = FALSE,
  kappa_1 = 1,
  kappa_2 = 1
)
}
\arguments{
\item{rankings}{A set of complete rankings, with one sample per column.
With n_assessors samples and n_items items, rankings is n_items x n_assessors.}

\item{obs_freq}{A vector of observation frequencies (weights) to apply to the rankings.}

\item{constraints}{List of lists of lists, returned from `generate_constraints`.}

\item{compute_options}{An object of class \code{"BayesMallowsComputeOptions"}
returned from \code{\link{set_compute_options}}.}

\item{cardinalities}{Used when metric equals \code{"footrule"} or
\code{"spearman"} for computing the partition function. Defaults to
\code{R_NilValue}.}

\item{logz_estimate}{Estimate of the log partition function.}

\item{metric}{The distance metric to use. One of \code{"spearman"},
\code{"footrule"}, \code{"kendall"}, \code{"cayley"}, or
\code{"hamming"}.}

\item{error_model}{Error model to use.}

\item{n_clusters}{Number of clusters. Defaults to 1.}

\item{alpha_init}{Initial value of alpha.}

\item{verbose}{Logical specifying whether to print out the progress of the
Metropolis-Hastings algorithm. If \code{TRUE}, a notification is printed every
1000th iteration.}

\item{kappa_1}{Hyperparameter for \eqn{theta} in the Bernoulli error model. Defaults to 1.0.}

\item{kappa_2}{Hyperparameter for \eqn{theta} in the Bernoulli error model. Defaults to 1.0.}

\item{lambda}{Parameter of the prior distribution.}

\item{alpha_max}{Maximum value of \code{alpha}, used for truncating the exponential prior distribution.}

\item{psi}{Hyperparameter for the Dirichlet prior distribution used in clustering.}
}
\description{
Worker function for computing the posterior distribution.
}
\keyword{internal}
