% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{run_mcmc}
\alias{run_mcmc}
\title{Worker function for computing the posterior distribtuion.}
\usage{
run_mcmc(R, nmc, cardinalities, metric = "footrule", L = 1L,
  sd_alpha = 0.5, alpha_init = 5, alpha_jump = 1L, lambda = 0.1)
}
\arguments{
\item{R}{A set of complete rankings, with one sample per column.
With N samples and n items, R is n x N.}

\item{nmc}{Number of Monte Carlo samples.}

\item{cardinalities}{Used when metric equals \code{"footrule"} or
\code{"spearman"} for computing the partition function. Otherwise,
please provided an arbitrary vector.}

\item{metric}{The distance metric to use. On of \code{"spearman"},
\code{"footrule"}, \code{"kendall"}, \code{"cayley"}, or
\code{"hamming"}.}

\item{L}{Leap-and-shift step size.}

\item{sd_alpha}{Standard deviation of proposal distribution for alpha.}

\item{alpha_init}{Initial value of alpha.}

\item{alpha_jump}{How many times should we sample \code{rho} between
each time we sample \code{alpha}. Setting \code{alpha_jump} to a high
number can significantly speed up computation time, since we then do not
have to do expensive computation of the partition function.}

\item{lambda}{Parameter of the prior distribution.}
}
\description{
Worker function for computing the posterior distribtuion.
}
