% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R, R/smc_mallows_deprecated.R
\name{smc_mallows_new_item_rank}
\alias{smc_mallows_new_item_rank}
\alias{smc_mallows_new_item_rank_alpha_fixed}
\title{SMC-Mallows new users rank}
\usage{
smc_mallows_new_item_rank(
  n_items,
  R_obs,
  metric,
  leap_size,
  N,
  Time,
  logz_estimate,
  mcmc_kernel_app,
  alpha = 0,
  alpha_prop_sd = 1,
  lambda = 1,
  alpha_max = 1,
  aug_method = "random",
  verbose = FALSE,
  alpha_fixed = FALSE
)

smc_mallows_new_item_rank_alpha_fixed(
  alpha,
  n_items,
  R_obs,
  metric,
  leap_size,
  N,
  Time,
  logz_estimate,
  mcmc_kernel_app,
  alpha_prop_sd,
  lambda,
  alpha_max,
  aug_method,
  verbose = FALSE
)
}
\arguments{
\item{n_items}{Integer is the number of items in a ranking}

\item{R_obs}{3D matrix of size n_assessors by n_items by Time containing a set of observed rankings of Time time steps}

\item{metric}{A character string specifying the distance metric to use in the
Bayesian Mallows Model. Available options are \code{"footrule"},
\code{"spearman"}, \code{"cayley"}, \code{"hamming"}, \code{"kendall"}, and
\code{"ulam"}.}

\item{leap_size}{leap_size Integer specifying the step size of the leap-and-shift
proposal distribution}

\item{N}{Integer specifying the number of particles}

\item{Time}{Integer specifying the number of time steps in the SMC algorithm}

\item{logz_estimate}{Estimate of the partition function, computed with
\code{\link{estimate_partition_function}} in the BayesMallow R package {estimate_partition_function}.}

\item{mcmc_kernel_app}{Integer value for the number of applications we apply the MCMC move kernel}

\item{alpha_prop_sd}{Numeric value of the standard deviation of the prior distribution for alpha}

\item{lambda}{Strictly positive numeric value specifying the rate parameter
of the truncated exponential prior distribution of alpha.}

\item{alpha_max}{Maximum value of alpha in the truncated exponential
prior distribution.}

\item{aug_method}{A character string specifying the approach for filling in the missing data, options are "pseudolikelihood" or "random"}

\item{verbose}{Logical specifying whether to print out the progress of the
SMC-Mallows algorithm. Defaults to \code{FALSE}.}

\item{alpha_fixed}{Logical indicating whether to sample \code{alpha} or not.}
}
\value{
a set of particles each containing the values of rho and alpha and the effective sample size (ESS) at each iteration of the SMC
algorithm as well as the set of augmented rankings at the final iteration.
}
\description{
Function to perform resample-move SMC algorithm where we receive a new item ranks from an existing user
at each time step. Each correction and augmentation is done by filling in the missing item ranks using pseudolikelihood augmentation.
}
\section{Functions}{
\itemize{
\item \code{smc_mallows_new_item_rank_alpha_fixed()}: Deprecated function for
\code{smc_mallows_new_item_rank_alpha_fixed}.

}}
\keyword{internal}
