% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updated_rankings_functions_for_waldir.R
\name{smc_mallows_new_item_rank_updated}
\alias{smc_mallows_new_item_rank_updated}
\title{SMC-Mallows new item rank updated}
\usage{
smc_mallows_new_item_rank_updated(
  n_items,
  R_obs,
  metric,
  leap_size,
  N,
  Time,
  logz_estimate,
  mcmc_kernel_app,
  alpha_prop_sd,
  lambda,
  alpha_max,
  aug_method,
  alpha_samples_init,
  rho_samples_init,
  aug_rankings_init
)
}
\arguments{
\item{n_items}{Integer is the number of items in a ranking.}

\item{R_obs}{3D matrix of size n_assessors by n_items by Time containing a set of observed rankings of Time steps.}

\item{metric}{A character string specifying the distance metric to use in the Bayesian Mallows Model. Available options are \code{"footrule"},
\code{"spearman"}, \code{"cayley"}, \code{"hamming"}, \code{"kendall"}, and \code{"ulam"}.}

\item{leap_size}{leap_size Integer specifying the step size of the leap-and-shift proposal distribution.}

\item{N}{Integer specifying the number of particles.}

\item{Time}{Integer specifying the number of time steps in the SMC algorithm.}

\item{logz_estimate}{Estimate of the partition function, computed with \code{\link{estimate_partition_function}} in the BayesMallow R package {estimate_partition_function}.}

\item{mcmc_kernel_app}{Integer value for the number of applications we apply the MCMC move kernel.}

\item{alpha_prop_sd}{Numeric value of the standard deviation of the prior distribution for alpha.}

\item{lambda}{Strictly positive numeric value specifying the rate parameter of the truncated exponential prior distribution of alpha.}

\item{alpha_max}{Maximum value of alpha in the truncated exponential prior distribution.}

\item{aug_method}{A character string specifying the approach for filling in the missing data, options are "pseudolikelihood" or "random".}

\item{alpha_samples_init}{A vector of size N by containing the initial particle set values of alpha.}

\item{rho_samples_init}{2D matrix of size N by n_items containing the initial particle set values of rho.}

\item{aug_rankings_init}{3D matrix of size n_assessors by n_items by N containing the initial particle set values of augmented rankings for R_obs.}
}
\value{
a 3d matrix containing: the samples of: rho, alpha and the augmented rankings, and the effective sample size at each iteration of the SMC algorithm.
}
\description{
Function to perform resample-move SMC algorithm where we receive a new item ranks from an existing user at each time step given an initial particle set obtained from MCMC or SMC. Each correction and augmentation is done by filling in the missing item ranks using pseudolikelihood augmentation.
}
