% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smc_post_processing_functions.R
\name{smc_processing}
\alias{smc_processing}
\title{SMC Processing}
\usage{
smc_processing(output, colnames = NULL)
}
\arguments{
\item{output}{a subset of an SMCMallows object (though technically any matrix will do)}

\item{colnames}{colnames}
}
\value{
A processed file of the SMCMallows class
}
\description{
SMC Processing
}
\seealso{
\code{\link{smc_mallows_new_item_rank}} and
\code{\link{smc_mallows_new_users}}, which are functions generating objects
of SMCMallows class.
}
\author{
Anja Stein
}
\keyword{internal}
