% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sushi_rankings}
\alias{sushi_rankings}
\title{Sushi Rankings}
\format{
An object of class \code{matrix} (inherits from \code{array}) with 5000 rows and 10 columns.
}
\usage{
sushi_rankings
}
\description{
Complete rankings of 10 types of sushi from
5000 assessors \insertCite{kamishima2003}{BayesMallows}.
}
\references{
\insertAllCited{}
}
\keyword{datasets}
