% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_mallows.R
\name{update_mallows}
\alias{update_mallows}
\alias{update_mallows.BayesMallows}
\alias{update_mallows.SMCMallows}
\title{Update a Bayesian Mallows model with new users}
\usage{
update_mallows(model, new_rankings, ...)

\method{update_mallows}{BayesMallows}(
  model,
  new_rankings,
  n_particles,
  augmentation = "pseudo",
  mcmc_steps = 5,
  ...
)

\method{update_mallows}{SMCMallows}(model, new_rankings, ...)
}
\arguments{
\item{model}{A model object.}

\item{new_rankings}{Matrix containing the new set of observed rankings of size
n_assessors by n_items.}

\item{...}{Optional arguments. Currently not used.}

\item{n_particles}{Integer specifying the number of particles.}

\item{augmentation}{One of "pseudo" and "uniform".}

\item{mcmc_steps}{Number of Metropolis-Hastings steps to apply in sequential
Monte Carlo.}
}
\value{
An updated model, of class "SMCMallows".
}
\description{
Update a Bayesian Mallows model with new users
}
\seealso{
Other modeling: 
\code{\link{compute_mallows_mixtures}()},
\code{\link{compute_mallows}()}
}
\concept{modeling}
